### 2020-07-28 #2531 @whedon
[PRE REVIEW]: cfdm: A Python reference implementation of the CF data model
**Submitting author:** @davidhassell (<a href="http://orcid.org/0000-0001-5106-7502">David Hassell</a>)
**Repository:** <a href="https://github.com/NCAS-CMS/cfdm" target ="_blank">https://github.com/NCAS-CMS/cfdm</a>
**Version:** 1.8.6.0
