### 2020-07-29 ee21eb2
Remove test suite outcome dependence on run dir
