+++ cfdm/read_write/netcdf/netcdfread.py
@@ -2072,7 +2072,7 @@ class NetCDFRead(IORead):
 
             # --------------------------------------------------------
             # Cell dimension can not be taken from the node_count
-            # variable (becuase it doesn't exist), so it has to be
+            # variable (because it doesn't exist), so it has to be
             # taken from one of the node_coordinate variables,
             # instead.
             # --------------------------------------------------------
@@ -2102,6 +2102,8 @@ class NetCDFRead(IORead):
         self.implementation.nc_set_sample_dimension(
             nodes_per_geometry, self._ncdim_abspath(node_dimension))
 
+        
+        print ( 'part_node_count=', part_node_count)
         if part_node_count is None:
             # --------------------------------------------------------
             # There is no part_count variable, i.e. cell has exactly
@@ -2109,14 +2111,18 @@ class NetCDFRead(IORead):
             #
             # => we can treat the nodes as a contiguous ragged array
             # --------------------------------------------------------
+            print ('111 self._set_ragged_contiguous_parameters')
             element_dimension = self._set_ragged_contiguous_parameters(
                 elements_per_instance=nodes_per_geometry,
                 sample_dimension=node_dimension,
                 element_dimension='node',
                 instance_dimension=geometry_dimension)
-
-            g['compression'][node_dimension]['netCDF_variables'] = (
-                parsed_node_coordinates[:])
+             
+#            g['compression'][node_dimension]['netCDF_variables'] = (
+#                parsed_node_coordinates[:])
+            g['compression'][node_dimension].setdefault(
+                'netCDF_variables', set()).update(parsed_node_coordinates)
+            print ("g['compression']", g['compression'])
         else:
             # --------------------------------------------------------
             # There is a part_count variable.
@@ -2139,7 +2145,7 @@ class NetCDFRead(IORead):
             total_number_of_parts = self.implementation.get_data_size(parts)
 
             parts_data = self.implementation.get_data(parts)
-#            print ('PARTS_DATA=',repr(parts_data), parts_data.array)
+
             nodes_per_geometry_data = self.implementation.get_data(
                 nodes_per_geometry)
 
@@ -2152,11 +2158,10 @@ class NetCDFRead(IORead):
                     nodes_per_geometry)):
                 n_nodes_in_this_cell = int(nodes_per_geometry_data[cell_no])
 
-                # Initiailize partial_node_count, a running count
-                # of how many nodes there are in this geometry
+                # Initialise partial_node_count, a running count of
+                # how many nodes there are in this geometry
                 n_nodes = 0
 
-#                print ('cell_no=', cell_no)
                 for k in range(i, total_number_of_parts):
                     index.data[k] = instance_index
                     n_nodes += int(parts_data[k])
@@ -2844,7 +2849,7 @@ class NetCDFRead(IORead):
         field_ncdimensions = self._ncdimensions(field_ncvar)
 
         field_groups = g['variable_groups'][field_ncvar]
-
+        print ('\n\n\n\n', field_ncvar, 'field_ncdimensions=',field_ncdimensions)
         for ncdim in field_ncdimensions:
             ncvar, method = self._find_coordinate_variable(field_ncvar,
                                                            field_groups,
@@ -3112,9 +3117,15 @@ class NetCDFRead(IORead):
                 # the data variable
                 geometry_dimension = geometry['geometry_dimension']
                 if geometry_dimension not in g['ncdim_to_axis']:
-                    logger.warning(
-                        "geometry['geometry_dimension'] is not in "
-                        "g['ncdim_to_axis']"
+                    ooo = self._get_geometry(field_ncvar, return_ncvar=1)
+                    print (field_ncvar, ooo, geometry)
[…]
