### 2020-07-29 #2531 @usethedata
@davidhassell I'm in the middle of a quarterly planning meeting this week.  I'll dig into your paper and repo over the weekend and start looking for reviewers.  Thanks for your submission and for your patience.
