+++ cfdm/read_write/netcdf/netcdfwrite.py
@@ -166,7 +166,7 @@ class NetCDFWrite(IOWrite):
                     # given role.
                     return ncdim
         # --- End: if
-#ppp
+
         if base in existing_names:
             counter = g.setdefault('count_' + base, 1)
 
@@ -183,7 +183,7 @@ class NetCDFWrite(IOWrite):
 
         if role and dimsize is not None:
             g['dimensions_with_role'].setdefault(role, []).append(ncvar)
-            
+
         return ncvar
 
     def _numpy_compressed(self, array):
@@ -674,9 +674,9 @@ class NetCDFWrite(IOWrite):
                 # netCDf variable name.
                 ncvar = self._create_netcdf_variable_name(coord,
                                                           default='coordinate')
-                
+
             ncdim = ncvar
-            
+
             # Create a new dimension
             unlimited = self.implementation.nc_is_unlimited_axis(f, axis)
 
@@ -706,12 +706,12 @@ class NetCDFWrite(IOWrite):
 #            # --- End: if
 #
 #            if g['group']:
-#                
+#
 #                _, groups = self._remove_group_structure(
 #                    ncvar, return_groups=True)
 #               ncvar = groups + self._remove_group_structure(ncdim)
 #            else:
-#                ncvar = ncdim    
+#                ncvar = ncdim
 
             self._write_dimension(ncdim, f, axis, unlimited=unlimited)
 
@@ -929,13 +929,12 @@ class NetCDFWrite(IOWrite):
 
             # assuming 1-d coord ...
             geometry_dimension = g['key_to_ncdims'][key][0]
-            
+
             geometry_id = (geometry_dimension, geometry_type)
             gc.setdefault(
                 geometry_id,
-                {'geometry_type': geometry_type,
-#                 'geometry_dimension': geometry_dimension,
-                }
+                {'geometry_type': geometry_type}
+                #                 'geometry_dimension': geometry_dimension,
             )
 
             # Nodes
@@ -1053,7 +1052,7 @@ class NetCDFWrite(IOWrite):
         _, geometry_container = gc.popitem()
 
         g['geometry_dimensions'].add(geometry_dimension)
-        
+
         return geometry_container
 
     def _already_in_file(self, variable, ncdims=None, ignore_type=False):
@@ -1286,7 +1285,7 @@ class NetCDFWrite(IOWrite):
                 ncvar = self._remove_group_structure(ncvar)
 
             ncvar = self._netcdf_name(ncvar)
-            
+
             # If no groups have been set on the bounds, then put the
             # bounds variable in the same group as its parent
             # coordinates
@@ -1294,7 +1293,7 @@ class NetCDFWrite(IOWrite):
             coord_groups = self._groups(coord_ncvar)
             if not bounds_groups and coord_groups:
                 ncvar = coord_groups + ncvar
-                
+
 #            for ncdim in ncdimensions:
 #                _, ncdim_groups = self._remove_group_structure(
 #                    ncdim,
@@ -1305,7 +1304,7 @@ class NetCDFWrite(IOWrite):
 #                        "with a dimension that is not in the same group or "
 #                        "a sub-group as the variable: {}".format(bounds, ncdim)
 #                    )
-            
+
             # Note that, in a field, bounds always have equal units to
             # their parent coordinate
 
@@ -2983,7 +2982,6 @@ class NetCDFWrite(IOWrite):
         for axis, domain_axis in sorted(domain_axes.items()):
             ncdim = self.implementation.nc_get_dimension(domain_axis,
                                                          default=None)
-            print ('ncdim =', ncdim)
 #            if ncdim is not None:
 #                ncdim = self._netcdf_name(ncdim)
 
[…]
