+++ cfdm/data/data.py
@@ -28,7 +28,7 @@ class Data(mixin.Container,
            core.Data):
     '''An orthogonal multidimensional array with masked values and units.
 
-    .. versionadded:: 1.7.0
+    .. versionadded:: (cfdm) 1.7.0
 
     '''
     def __init__(self, array=None, units=None, calendar=None,
@@ -107,7 +107,7 @@ class Data(mixin.Container,
             Apply this mask to the data given by the *array*
             parameter. By default, or if *mask* is `None`, no mask is
             applied. May be any scalar or array-like object (such as a
-            `numpy` array or `Data` instance) that is scalar or has
+            `numpy` array or `{{class}}` instance) that is scalar or has
             the same shape as *array*. Masking will be carried out
             where mask elements evaluate to `True`.
 
@@ -153,7 +153,7 @@ class Data(mixin.Container,
     def __array__(self, *dtype):
         '''The numpy array interface.
 
-    .. versionadded:: 1.7.0
+    .. versionadded:: (cfdm) 1.7.0
 
     :Parameters:
 
@@ -168,13 +168,13 @@ class Data(mixin.Container,
     **Examples:**
 
     >>> import numpy
-    >>> d = Data([1, 2, 3])
+    >>> d = {{package}}.{{class}}([1, 2, 3])
     >>> a = numpy.array(d)
     >>> print(type(a))
     <type 'numpy.ndarray'>
     >>> a[0] = -99
     >>> d
-    <Data(3): [1, 2, 3]>
+    <{{repr}}Data(3): [1, 2, 3]>
     >>> b = numpy.array(d, float)
     >>> print(b)
     [ 1.  2.  3.]
@@ -220,19 +220,19 @@ class Data(mixin.Container,
       the same behaviour as indexing on a Variable object of the
       netCDF4 package.
 
-    .. versionadded:: 1.7.0
+    .. versionadded:: (cfdm) 1.7.0
 
     .. seealso:: `__setitem__`, `_parse_indices`
 
     :Returns:
 
-        `Data`
+        `{{class}}`
             The subspace of the data.
 
     **Examples:**
 
     >>> import numpy
-    >>> d = Data(numpy.arange(100, 190).reshape(1, 10, 9))
+    >>> d = {{package}}.{{class}}(numpy.arange(100, 190).reshape(1, 10, 9))
     >>> d.shape
     (1, 10, 9)
     >>> d[:, :, 1].shape
@@ -285,25 +285,25 @@ class Data(mixin.Container,
 
     **Examples:**
 
-    >>> d = Data([1, 2, 3], 'metres')
+    >>> d = {{package}}.{{class}}([1, 2, 3], 'metres')
     >>> for e in d:
-    ...    print repr(e)
+    ...    print(repr(e))
     ...
     1
     2
     3
 
-    >>> d = Data([[1, 2], [4, 5]], 'metres')
+    >>> d = {{package}}.{{class}}([[1, 2], [4, 5]], 'metres')
     >>> for e in d:
-    ...    print repr(e)
+    ...    print(repr(e))
     ...
-    <CF Data: [1, 2] metres>
-    <CF Data: [4, 5] metres>
+    <{{repr}}Data: [1, 2] metres>
+    <{{repr}}Data: [4, 5] metres>
 
-    >>> d = Data(34, 'metres')
+    >>> d = {{package}}.{{class}}(34, 'metres')
     >>> for e in d:
-    ...     print repr(e)
-    ..
+    ...     print(repr(e))
+    ...
     TypeError: iteration over a 0-d Data
 
         '''
[…]
