### 2020-07-30 f7e55cb
fixed inheritance, renamed abstract -> mixin
