+++ cfdm/core/abstract/__init__.py
@@ -1,10 +1,0 @@
-from .rewritedocstring import RewriteDocstringMeta
-
-from .container                   import Container
-from .properties                  import Properties
-from .propertiesdata              import PropertiesData
-from .propertiesdatabounds        import PropertiesDataBounds
-from .coordinate                  import Coordinate
-
-from .parameters                  import Parameters
-from .parametersdomainancillaries import ParametersDomainAncillaries
