+++ cfdm/core/meta/__init__.py
@@ -1,0 +1,1 @@
+from .rewritedocstring import RewriteDocstringMeta
