+++ cfdm/core/domain.py
@@ -1,10 +1,9 @@
-from . import abstract
 from . import mixin
 
 from . import Constructs
 
 
-class Domain(mixin.ConstructAccess, abstract.Container):
+class Domain(mixin.ConstructAccess, mixin.Container):
     '''A domain of the CF data model.
 
     The domain represents a set of discrete "locations" in what
