+++ cfdm/core/mixin/__init__.py
@@ -1,5 +1,3 @@
-#from .rewritedocstring import RewriteDocstringMeta
-
 from .constructaccess import ConstructAccess
 
 from .container                   import Container
