+++ cfdm/core/mixin/coordinate.py
@@ -2,10 +2,15 @@ import abc
 
 from . import PropertiesDataBounds
 
+# --------------------------------------------------------------------
+# See cfdm.core.mixin.container.__docstring_substitution__ for
+# {{...}}  docstring substitutions
+# --------------------------------------------------------------------
 
-class Coordinate(PropertiesDataBounds): #, metaclass=abc.ABCMeta):
-    '''Abstract base class for dimension and auxiliary coordinate
-    constructs of the CF data model.
+
+class Coordinate(PropertiesDataBounds):
+    '''Mixin class for dimension and auxiliary coordinate constructs of
+    the CF data model.
 
     .. versionadded:: 1.7.0
 
