+++ cfdm/mixin/netcdf.py
@@ -1,12 +1,11 @@
 from copy import deepcopy
 
-#from ..core import abstract
 
 class DeprecationError(Exception):
     pass
 
 
-class NetCDF: #(abstract.ZeroContainer):
+class NetCDF:
     '''Mixin class for storing simple netCDF elements.
 
     .. versionadded:: 1.7.0

[…]
