+++ cfdm/test/test_docstring.py
@@ -1,0 +1,137 @@
+import datetime
+import unittest
+
+import cfdm
+
+
+class DocstringTest(unittest.TestCase):
+    def setUp(self):
+        self.package = 'cfdm'
+        self.repr = ''
+        self.subclasses_of_Container = (
+            cfdm.Field,
+            cfdm.AuxiliaryCoordinate,
+            cfdm.DimensionCoordinate,
+            cfdm.DomainAncillary,
+            cfdm.FieldAncillary,
+            cfdm.CellMeasure,
+            cfdm.DomainAxis,
+            cfdm.CoordinateReference,
+            cfdm.CellMethod,
+
+            cfdm.NodeCountProperties,
+            cfdm.PartNodeCountProperties,
+            cfdm.Bounds,
+            cfdm.InteriorRing,
+            cfdm.List,
+            cfdm.Index,
+            cfdm.Count,
+
+            cfdm.Data,
+            cfdm.NetCDFArray,
+            cfdm.NumpyArray,
+            cfdm.GatheredArray,
+            cfdm.RaggedContiguousArray,
+            cfdm.RaggedIndexedArray,
+            cfdm.RaggedIndexedContiguousArray,
+        )
+        self.subclasses_of_Properties = (
+            cfdm.Field,
+            cfdm.AuxiliaryCoordinate,
+            cfdm.DimensionCoordinate,
+            cfdm.DomainAncillary,
+            cfdm.FieldAncillary,
+            cfdm.CellMeasure,
+            cfdm.NodeCountProperties,
+            cfdm.PartNodeCountProperties,
+            cfdm.Bounds,
+            cfdm.InteriorRing,
+            cfdm.List,
+            cfdm.Index,
+            cfdm.Count,
+        )
+        self.subclasses_of_PropertiesData = (
+            cfdm.Field,
+            cfdm.AuxiliaryCoordinate,
+            cfdm.DimensionCoordinate,
+            cfdm.DomainAncillary,
+            cfdm.FieldAncillary,
+            cfdm.CellMeasure,
+            cfdm.Bounds,
+            cfdm.InteriorRing,
+            cfdm.List,
+            cfdm.Index,
+            cfdm.Count,
+        )
+        self.subclasses_of_PropertiesDataBounds = (
+            cfdm.AuxiliaryCoordinate,
+            cfdm.DimensionCoordinate,
+            cfdm.DomainAncillary,
+        )
+
+    def test_docstring(self):
+        # Test that all {{ occurences have been substituted
+        for klass in self.subclasses_of_Container:
+            for x in (klass, klass()):
+                for name in dir(x):
+                    if name.startswith('__'):
+                        continue
+                    
+                    f = getattr(klass, name)
+                    if not hasattr(f, '__doc__'):
+                        continue
+                    
+                    self.assertIsNotNone(f.__doc__,
+                                         '{}, {}, {}, {}, {}'.format(klass, name, f, dir(f), f.__doc__))
+                    
+                    self.assertNotIn('{{', f.__doc__,
+                                     '{}, {}, {}'.format(klass, name, f))
+
+    def test_docstring_package(self):
+        string = '>>> f = {}.'.format(self.package)
+        for klass in self.subclasses_of_Container:
+            for x in (klass, klass()):
+                self.assertIn(string, x._has_component.__doc__, klass)
+
+        string = '>>> f = {}.'.format(self.package)
+        for klass in self.subclasses_of_Properties:
+            for x in (klass, klass()):
+                self.assertIn(string, x.clear_properties.__doc__, klass)
+
+    def test_docstring_class(self):
+        for klass in self.subclasses_of_Properties:
+            string = '>>> f = {}.{}'.format(self.package, klass.__name__)
[…]
