### 2020-07-30 ac92826
Convert all anomalous docstring cases of 'cf.foo' to 'cfdm.foo'
