### 2020-07-31 ec51dec
Lots of print statements
