### 2020-08-03 51bbf92
Update test_read_write.py
removed test restrictions
