### 2020-08-03 e9c0e9a
Tests: update deprecated & insecure mktemp to mkstemp
