+++ cfdm/test/test_Field.py
@@ -12,7 +12,7 @@ import numpy
 import cfdm
 
 n_tmpfiles = 1
-tmpfiles = [tempfile.mktemp('_test_Field.nc', dir=os.getcwd())
+tmpfiles = [tempfile.mkstemp('_test_Field.nc', dir=os.getcwd())[1]
             for i in range(n_tmpfiles)]
 (tmpfile,) = tmpfiles
 
