+++ cfdm/test/test_netCDF.py
@@ -9,7 +9,7 @@ import cfdm
 
 
 n_tmpfiles = 3
-tmpfiles = [tempfile.mktemp('_test_netCDF.nc', dir=os.getcwd())
+tmpfiles = [tempfile.mkstemp('_test_netCDF.nc', dir=os.getcwd())[1]
             for i in range(n_tmpfiles)]
 (
     tempfile1,
