+++ cfdm/read_write/netcdf/netcdfwrite.py
@@ -5,6 +5,8 @@ import re
 
 from distutils.version import LooseVersion
 
+from pprint import (pformat, pprint)
+
 import numpy
 import netCDF4
 
@@ -1364,16 +1366,19 @@ class NetCDFWrite(IOWrite):
         `dict`
 
         '''
+        print ('      coord_ncdimensions=', coord_ncdimensions)
         out = {}
 
         g = self.write_vars
 
         bounds = self.implementation.get_bounds(coord, None)
         if bounds is None:
+            # This coordinate construct has no bounds
             return {}
 
         data = self.implementation.get_data(bounds, None)
         if data is None:
+            # The bounds have no data
             return {}
 
         # Still here? Then this coordinate has a nodes attribute
@@ -1403,6 +1408,8 @@ class NetCDFWrite(IOWrite):
         # Find the base of the netCDF part dimension name
         size = self.implementation.get_data_size(nodes)
         ncdim = self._get_node_ncdimension(nodes, default='node')
+        print ('   node ncdim =', ncdim, repr(bounds),
+               bounds.nc_get_variable() , bounds.nc_get_dimension())
         ncdim = self._netcdf_name(ncdim, dimsize=size, role='node')
 
         if self._already_in_file(nodes, (ncdim,)):
@@ -1715,14 +1722,20 @@ class NetCDFWrite(IOWrite):
         return groups
 
     def _get_node_ncdimension(self, bounds, default=None):
-
         '''TODO
 
     .. versionadded:: 1.8.0
 
+    :Parameters:
+
+        bounds: Bounds component
+
+        default: optional
+
     :Returns:
 
-        TODO
+        The netCDF dimension name, or else the value of the *default*
+        parameter.
 
         '''
         ncdim = self.implementation.nc_get_dimension(bounds, default=None)
@@ -1735,7 +1748,7 @@ class NetCDFWrite(IOWrite):
 
             return ncdim
 
-        # Retrun the default
+        # Return the default
         return default
 
     def _write_part_node_count(self, coord, bounds, encodings):
@@ -2010,19 +2023,15 @@ class NetCDFWrite(IOWrite):
 
         key: `str`
 
-        coord: `Coordinate`
+        coord: Coordinate construct
 
         coordinates: `list`
 
     :Returns:
 
-        coordinates: `list`
+        `list`
             The list of netCDF auxiliary coordinate names updated in
             place.
-
-    **Examples:**
-
-    >>> coordinates = _write_auxiliary_coordinate(f, 'aux2', coordinates)
 
         '''
         g = self.write_vars
@@ -2821,9 +2830,9 @@ class NetCDFWrite(IOWrite):
 
     :Parameters:
 
-        f : Field construct
+        f: Field construct
 
-        add_to_seen : bool, optional
+        add_to_seen: bool, optional
 
         allow_data_insert_dimension: `bool`, optional
 
@@ -2985,6 +2994,8 @@ class NetCDFWrite(IOWrite):
 #            if ncdim is not None:
 #                ncdim = self._netcdf_name(ncdim)
 
+            print ('\n\nncdim = ', ncdim)
+
             found_dimension_coordinate = False
             for key, dim_coord in dimension_coordinates.items():
                 if (self.implementation.get_construct_data_axes(f, key)
@@ -3039,7 +3050,7 @@ class NetCDFWrite(IOWrite):
                 found_dimension_coordinate = True
                 break
             # --- End: for
-
+            print (  '  comment =', f.get_property('comment', None))  
             if not found_dimension_coordinate:
                 # ----------------------------------------------------
                 # There is NO dimension coordinate for this axis
@@ -3047,7 +3058,7 @@ class NetCDFWrite(IOWrite):
                 spanning_constructs = (
[…]
