+++ cfdm/test/test_read_write.py
@@ -411,33 +411,44 @@ class read_writeTest(unittest.TestCase):
                 "{!r}, {!r}".format(
                     g.get_property('Conventions'), Conventions))
 
-#    def test_read_write_duplicate_names(self):
-#        if self.test_only and inspect.stack()[0][3] not in self.test_only:
-#            return
-#
-#        a = []
-#        for filename in (
-#                'geometry_1.nc',
-#                'geometry_2.nc',
-#                'geometry_3.nc',
-#                'geometry_4.nc',
-#                'geometry_interior_ring_2.nc',
-#                'geometry_interior_ring.nc',
-#        ):
-#            a.extend(cfdm.read(filename))
-#        print (a)
-#        print (a[0].dump())
-#        print ('THE WRITE')
-#        tmpfile = 'delme.nc'
-#        cfdm.write(a, tmpfile, verbose=-1)
-#        print ('\n\n\n\n FINAL READ \n\n\n\n\n')
-#        f = cfdm.read(tmpfile, verbose=-1)
-#        print ('WWWWWWWW')
+    def test_read_write_duplicate_names(self):
+        if self.test_only and inspect.stack()[0][3] not in self.test_only:
+            return
+
+        a = []
+        for filename in (
+                'geometry_1.nc',
+                'geometry_2.nc',
+                'geometry_3.nc',
+                'geometry_4.nc',
+                'geometry_interior_ring_2.nc',
+                'geometry_interior_ring.nc',
+        ):
+            
+            a.extend(cfdm.read(filename))
+            print ('ppp', len(a))
+            
+        print (a)
+        print (a[0].dump())
+        for n, f in enumerate(a):
+            f.set_property('test_id', str(n))
+            if n == 4:
+                f.dump()
+                print (f.construct('latitude').nc_get_variable())
+                print (f.construct('latitude').bounds.nc_get_variable())
+                print (f.construct('latitude').bounds.nc_get_dimension())
+                
+        print ('THE WRITE')
+        tmpfile = 'delme.nc'
+        cfdm.write(a, tmpfile, verbose=1)
+        print ('\n\n\n\n FINAL READ \n\n\n\n\n')
+        f = cfdm.read(tmpfile, verbose=-1)
+        print ('WWWWWWWW')
 # --- End: class
 
 
 if __name__ == "__main__":
     print('Run date:', datetime.datetime.now())
-    cfdm.environment(display=False)
+    cfdm.environment()
     print('')
     unittest.main(verbosity=2)
