+++ cfdm/read_write/netcdf/netcdfread.py
@@ -1865,9 +1865,9 @@ class NetCDFRead(IORead):
         g = self.read_vars
 
         logger.debug(
-            "    Pre-processing indexed and contiguous compression"
+            "    Pre-processing indexed and contiguous compression "
+            "for instance dimension: {}".format(instance_dimension)
         )  # pragma: no cover
-        logger.debug(g['compression'])  # pragma: no cover
 
         profile_dimension = g['compression'][sample_dimension][
             'ragged_contiguous']['profile_dimension']
@@ -1911,7 +1911,7 @@ class NetCDFRead(IORead):
         }
 
         logger.debug(
-            "    Created g['compression'][{!r}]"
+            "    Created read_vars['compression'][{!r}]"
             "['ragged_indexed_contiguous']".format(
                   sample_dimension)
         )  # pragma: no cover
@@ -1923,12 +1923,13 @@ class NetCDFRead(IORead):
                     'ragged_indexed_contiguous']['implied_ncdimensions']
             )
         )  # pragma: no cover
-        logger.debug(
-            "    Removing "
-            "read_vars['compression'][{!r}]['ragged_contiguous']".format(
-                sample_dimension))  # pragma: no cover
 
         del g['compression'][sample_dimension]['ragged_contiguous']
+
+        logger.debug(
+            "    Removed "
+            "read_vars['compression'][{!r}]['ragged_contiguous']".format(
+                sample_dimension))  # pragma: no cover
 
     def _parse_geometry(self, parent_ncvar, attributes):
         '''TODO
@@ -1967,14 +1968,6 @@ class NetCDFRead(IORead):
 
         geometry_ncvar = parsed_geometry[0]
 
-        logger.info(
-            "    Geometry container = {!r}".format(geometry_ncvar)
-        )  # pragma: no cover
-        logger.debug(
-            "        netCDF attributes: {}".format(
-                attributes[geometry_ncvar])
-        )  # pragma: no cover
-
         if geometry_ncvar in g['geometries']:
             # We've already parsed this geometry container, so record
             # the fact that this parent netCDF variable has this
@@ -1982,6 +1975,22 @@ class NetCDFRead(IORead):
             g['variable_geometry'][parent_ncvar] = geometry_ncvar
             return
 
+        logger.info(
+            "    Geometry container = {!r}".format(geometry_ncvar)
+        )  # pragma: no cover
+        logger.debug(
+            "        netCDF attributes: {}".format(
+                 pformat(attributes[geometry_ncvar], indent=12)
+            )
+        )  # pragma: no cover
+        
+#        logger.debug("    0 +++++++++++ Compression read vars:")  # pragma: no# cover
+#        logger.debug(
+#            "        read_vars['compression'] =\n" +
+#            pformat(g['compression'], indent=12)
+#        )  # pragma: no cover
+
+        
         geometry_type = attributes[geometry_ncvar].get('geometry_type')
 
         g['geometries'][geometry_ncvar] = {'geometry_type': geometry_type}
@@ -2114,14 +2123,9 @@ class NetCDFRead(IORead):
                 sample_dimension=node_dimension,
                 element_dimension='node',
                 instance_dimension=geometry_dimension)
-
-#            g['compression'][node_dimension]['netCDF_variables'] = (
-#                parsed_node_coordinates[:])
-            g['compression'][node_dimension].setdefault(
-                'netCDF_variables', set()).update(parsed_node_coordinates)
         else:
             # --------------------------------------------------------
-            # There is a part_count variable.
+            # There is a part node count variable.
             #
             # => we must treat the nodes as an indexed contiguous
             # ragged array
@@ -2173,13 +2177,17 @@ class NetCDFRead(IORead):
                 element_dimension='node',
                 instance_dimension=part_dimension)
 
[…]
