### 2020-08-04 6fb0aa0
NetCDF read: fix bug in _check_auxiliary_scalar_coordinate (typo)
