### 2020-08-04 7f64a07
multiple geometries now read/write OK. All unit tests pass
