+++ cfdm/read_write/netcdf/netcdfwrite.py
@@ -735,7 +735,7 @@ class NetCDFWrite(IOWrite):
         g['key_to_ncvar'][key] = ncvar
         g['key_to_ncdims'][key] = ncdimensions
 
-        g['axis_to_ncdim'][axis] = ncdim
+        g['axis_to_ncdim'][axis] = seen[id(coord)]['ncdims'][0]
 
         return ncvar
 
@@ -1366,7 +1366,6 @@ class NetCDFWrite(IOWrite):
         `dict`
 
         '''
-        print ('      coord_ncdimensions=', coord_ncdimensions)
         out = {}
 
         g = self.write_vars
@@ -1408,22 +1407,39 @@ class NetCDFWrite(IOWrite):
         # Find the base of the netCDF part dimension name
         size = self.implementation.get_data_size(nodes)
         ncdim = self._get_node_ncdimension(nodes, default='node')
-        print ('   node ncdim =', ncdim, repr(bounds),
-               bounds.nc_get_variable() , bounds.nc_get_dimension())
         ncdim = self._netcdf_name(ncdim, dimsize=size, role='node')
 
+        create = True
         if self._already_in_file(nodes, (ncdim,)):
             # This node coordinates variable has been previously
             # created, so no need to do so again.
             ncvar = g['seen'][id(nodes)]['ncvar']
 
-            # We need to log the original Bounds variable as being in
-            # the file, too. This is so that the geometry container
-            # variable can be created later on.
-            g['seen'][id(bounds)] = {'ncvar': ncvar,
-                                     'variable': bounds,
-                                     'ncdims': None}
-        else:
+            geometry_dimension = (
+                g['geometry_encoding'][ncvar]['geometry_dimension']
+            )
+
+            if geometry_dimension == coord_ncdimensions[0]:
+                # The node coordiante variable already exists, and the
+                # corresponding encoding variables span the correct
+                # dimension.
+                create = False
+
+                # We need to log the original Bounds variable as being
+                # in the file, too. This is so that the geometry
+                # container variable can be created later on.
+                g['seen'][id(bounds)] = {'ncvar': ncvar,
+                                         'variable': bounds,
+                                         'ncdims': None}
+            else:
+                # The node coordiante variable already exists, but the
+                # corresponding encoding variables span the wrong
+                # dimension => we have to create a new node
+                # coordinates variable.
+                create = True
+        # --- End: if
+
+        if create:
             # This node coordinates variable has not been previously
             # created, so create it now.
             ncdim_to_size = g['ncdim_to_size']
@@ -1469,15 +1485,18 @@ class NetCDFWrite(IOWrite):
 
             encodings = {}
 
-            _ = self._write_node_count(coord, bounds,
-                                       coord_ncdimensions, encodings)
-            encodings.update(_)
+            nc_encodings = self._write_node_count(coord, bounds,
+                                                  coord_ncdimensions,
+                                                  encodings)
+            encodings.update(nc_encodings)
 
-            _ = self._write_part_node_count(coord, bounds, encodings)
-            encodings.update(_)
+            pnc_encodings = self._write_part_node_count(coord, bounds,
+                                                        encodings)
+            encodings.update(pnc_encodings)
 
-            _ = self._write_interior_ring(coord, bounds, encodings)
-            encodings.update(_)
+            ir_encodings = self._write_interior_ring(coord, bounds,
+                                                     encodings)
+            encodings.update(ir_encodings)
 
             g['geometry_encoding'][ncvar] = encodings
 
@@ -1509,6 +1528,7 @@ class NetCDFWrite(IOWrite):
         bounds:
 
         coord_ncdimensions: sequence of `str`
+            The netCDF instance dimension
 
         encodings: `dict`
             Ignored.
[…]
