### 2020-08-04 1f412e3
fix for Python 3.5.9 test failures
