### 2020-08-05 c570c3e
wrapped methods & staticmethods
