+++ cfdm/core/meta/rewritedocstring.py
@@ -1,5 +1,8 @@
 import inspect
 import re
+import types
+
+#from functools import partial
 
 
 class RewriteDocstringMeta(type):
@@ -26,7 +29,7 @@ class RewriteDocstringMeta(type):
     # Define the "plus class" regular expression
     _plus_class_regex = re.compile('{{\+(\w.*?)}}')
 
-    def __new__(cls, name, parents, attrs):
+    def __new__(cls, class_name, parents, attrs):
         '''TODO
 
         '''
@@ -49,56 +52,95 @@ class RewriteDocstringMeta(type):
         if class_docstring_rewrite is not None:
             docstring_rewrite.update(class_docstring_rewrite(None))
 
+        for key in docstring_rewrite:
+            if key in ('{{class}}', '{{package}}') or key.startswith('{{+'):
+                raise ValueError(
+                    "Can't use {0!r} as a user-defined docstring substitution."
+                    "\nRemove the {0!r} key from the dictionary "
+                    "returned by the appropriate __docstring_substitution__ "
+                    "method.".format(key)
+                )
+        # --- End: for
+        
         for attr_name, attr in attrs.items():
             # Skip special methods
             if attr_name.startswith('__'):
                 continue
 
-            if name == 'Field':
-                print ('  ', attr_name)
             # Skip methods without docstrings
             if not hasattr(attr, '__doc__'):
                 continue
 
             # @property or normal method
-            if hasattr(attr, '__call__') or hasattr(attr, 'fget'):
-                # Update docstring
-                RewriteDocstringMeta._docstring_update(name, attr,
+            if hasattr(attr, 'fget'):
+                # Note that here inspect.isroutine(attr) is False for
+                # @property methods (but this is not the case for
+                # properties in the parent classes).
+
+                RewriteDocstringMeta._docstring_update(class_name, attr,
                                                        attr_name,
                                                        attrs['__module__'],
                                                        docstring_rewrite)
                 continue
 
-            # classmethod
-            if isinstance(attr, classmethod):
-                f = getattr(attr, '__func__')
+            if not inspect.isroutine(attr):
+                continue
 
+            is_classmethod = False
+            is_staticmethod = False
+            
+            # Find out if the method is a classmethod (the
+            # inspect.ismethod technique doesn't work for this class)
+            if isinstance(attr, classmethod):
+                is_classmethod = True
+            elif isinstance(attr, staticmethod):
+                is_staticmethod = True
+                
+            # Skip non-callables
+#            if not (is_classmethod or is_staticmethod) and not hasattr(attr, '#__call__'):
+#                continue
+
+            is_wrapped = hasattr(attr, '__wrapped__')
+            if is_wrapped:
+                wrapper = attr
+                attr = attr.__wrapped__
+
+            if is_classmethod or is_staticmethod:
+                f = getattr(attr, '__func__', attr)
+                
                 # Copy the method
                 attr = type(f)(f.__code__, f.__globals__,
                                f.__name__, f.__defaults__,
                                f.__closure__)
+            
+            # Update docstring
+            RewriteDocstringMeta._docstring_update(class_name, attr,
+                                                   attr_name,
+                                                   attrs['__module__'],
+                                                   docstring_rewrite)
 
-                # Update docstring
-                RewriteDocstringMeta._docstring_update(name, attr,
-                                                       attr_name,
-                                                       attrs['__module__'],
-                                                       docstring_rewrite)
-
+            # Redecorate
+            if is_classmethod:
                 attrs[attr_name] = classmethod(attr)
+
+            if is_staticmethod:
+                attrs[attr_name] = staticmethod(attr)
+                
+            if is_wrapped:
[…]
