### 2020-08-05 39ad002
Catch potentially obscure error with bad args to example_field()
