+++ cfdm/test/test_netCDF.py
@@ -727,6 +727,6 @@ class NetCDFTest(unittest.TestCase):
 
 if __name__ == '__main__':
     print('Run date:', datetime.datetime.now())
-    cfdm.environment(display=False)
+    cfdm.environment()
     print('')
     unittest.main(verbosity=2)
