+++ cfdm/core/mixin/__init__.py
@@ -1,6 +1,5 @@
 from .constructaccess import ConstructAccess
 
-from .container                   import Container
 from .properties                  import Properties
 from .propertiesdata              import PropertiesData
 from .propertiesdatabounds        import PropertiesDataBounds
