+++ cfdm/test/test_gathering.py
@@ -228,6 +228,6 @@ class GatheredTest(unittest.TestCase):
 
 if __name__ == '__main__':
     print('Run date:', datetime.datetime.utcnow())
-    cfdm.environment(display=False)
+    cfdm.environment()
     print()
     unittest.main(verbosity=2)
