+++ cfdm/test/test_external.py
@@ -189,6 +189,6 @@ class ExternalVariableTest(unittest.TestCase):
 
 if __name__ == '__main__':
     print('Run date:', datetime.datetime.utcnow())
-    cfdm.environment(display=False)
+    cfdm.environment()
     print()
     unittest.main(verbosity=2)
