+++ cfdm/core/meta/rewritedocstring.py
@@ -1,8 +1,5 @@
 import inspect
 import re
-import types
-
-#from functools import partial
 
 
 class RewriteDocstringMeta(type):
@@ -17,8 +14,9 @@ class RewriteDocstringMeta(type):
     wrong docstring. Instead, we must actually copy the functions, and
     then modify the docstring.
 
-    Special treatment is given to methods decorated with @property and
-    @classmethod. (Should other decorations be considered?)
+    Special treatment is given to methods decorated with @property,
+    @staticmethod and @classmethod, as well as user-defined
+    decorations.
 
     Based on
     http://www.jesshamrick.com/2013/04/17/rewriting-python-docstrings-with-a-metaclass/
@@ -26,7 +24,9 @@ class RewriteDocstringMeta(type):
     .. versionadded:: (cfdm) 1.8.7
 
     '''
-    # Define the "plus class" regular expression
+    # Define the "plus class" regular expression.
+    #
+    # E.g. matches: {{+Data}} with Data as the grouped match
     _plus_class_regex = re.compile('{{\+(\w.*?)}}')
 
     def __new__(cls, class_name, parents, attrs):
@@ -61,7 +61,9 @@ class RewriteDocstringMeta(type):
                     "method.".format(key)
                 )
         # --- End: for
-        
+
+        module = attrs['__module__']
+
         for attr_name, attr in attrs.items():
             # Skip special methods
             if attr_name.startswith('__'):
@@ -76,29 +78,26 @@ class RewriteDocstringMeta(type):
                 # Note that here inspect.isroutine(attr) is False for
                 # @property methods (but this is not the case for
                 # properties in the parent classes).
-
                 RewriteDocstringMeta._docstring_update(class_name, attr,
                                                        attr_name,
-                                                       attrs['__module__'],
+                                                       module,
                                                        docstring_rewrite)
                 continue
 
+            # Still here?
             if not inspect.isroutine(attr):
                 continue
 
+            # Still here?
             is_classmethod = False
             is_staticmethod = False
-            
+
             # Find out if the method is a classmethod (the
             # inspect.ismethod technique doesn't work for this class)
             if isinstance(attr, classmethod):
                 is_classmethod = True
             elif isinstance(attr, staticmethod):
                 is_staticmethod = True
-                
-            # Skip non-callables
-#            if not (is_classmethod or is_staticmethod) and not hasattr(attr, '#__call__'):
-#                continue
 
             is_wrapped = hasattr(attr, '__wrapped__')
             if is_wrapped:
@@ -107,16 +106,15 @@ class RewriteDocstringMeta(type):
 
             if is_classmethod or is_staticmethod:
                 f = getattr(attr, '__func__', attr)
-                
+
                 # Copy the method
                 attr = type(f)(f.__code__, f.__globals__,
                                f.__name__, f.__defaults__,
                                f.__closure__)
-            
+
             # Update docstring
             RewriteDocstringMeta._docstring_update(class_name, attr,
-                                                   attr_name,
-                                                   attrs['__module__'],
+                                                   attr_name, module,
                                                    docstring_rewrite)
 
             # Redecorate
@@ -125,39 +123,47 @@ class RewriteDocstringMeta(type):
 
             if is_staticmethod:
                 attrs[attr_name] = staticmethod(attr)
-                
+
             if is_wrapped:
                 wrapper.__doc__ = attr.__doc__
                 wrapper.__wrapped__ = attr
                 attrs[attr_name] = wrapper
         # --- End: for
-
+        ok = False
+        if class_name == 'DimensionCoordinate' and module.startswith('cf.'):
+            ok = True
+            print (module)
+            print (parents)
         for parent in parents:
             for attr_name in dir(parent):
-                
+                 
                 if attr_name in attrs:
[…]
