+++ cfdm/test/test_dsg.py
@@ -302,6 +302,6 @@ class DSGTest(unittest.TestCase):
 
 if __name__ == '__main__':
     print('Run date:', datetime.datetime.utcnow())
-    cfdm.environment(display=False)
+    cfdm.environment()
     print()
     unittest.main(verbosity=2)

[…]
