### 2020-08-07 77029bb
Reformulate uses of mutable default arguments across codebase
