### 2020-08-07 86b2ec2
Tidy test_read_write: reduce duplication of netCDF formats
