### 2020-08-07 89b94ee
Minimal unit test to codify nature of new netCDF append mode
