### 2020-08-07 e070c5f
Add 'mode' as a kwarg to the write() functions & document it
