+++ cfdm/read_write/netcdf/netcdfwrite.py
@@ -4339,6 +4339,7 @@ class NetCDFWrite(IOWrite):
         fields,
         filename,
         fmt="NETCDF4",
+        mode="w",
         overwrite=True,
         global_attributes=None,
         variable_attributes=None,
@@ -4388,6 +4389,22 @@ class NetCDFWrite(IOWrite):
                 The output CF-netCDF file.
 
                 See `cfdm.write` for details.
+
+            mode: `str`, optional
+                Specify the mode of write access for the output file. One of:
+                =======  =================================================
+                *mode*   Description
+                =======  =================================================
+                ``'w'``  Open a new file for writing to. If it exists and
+                         *overwrite* is True then the file is deleted
+                         prior to being recreated.
+                ``'a'``  Open an existing file for appending new
+                         information to. The new information will be
+                         incorporated whilst the original contents of the
+                         file will be preserved.
+                =======  =================================================
+                By default the file is opened with write access mode
+                ``'w'``.
 
             overwrite: bool, optional
                 If False then raise an exception if the output file
