### 2020-08-07 1e06308
Correct some typos (including one user-facing) in netcdfwrite
