+++ cfdm/data/mixin/__init__.py
@@ -1,2 +1,4 @@
-from .ragged import (RaggedContiguous,
-                     RaggedIndexed)
+from .arraymixin import ArrayMixin
+from .compressedarray import CompressedArray
+from .raggedcontiguous import RaggedContiguous
+from .raggedindexed import RaggedIndexed
