+++ cfdm/core/functions.py
@@ -14,7 +14,7 @@ from . import (__version__,
 def environment(display=True, paths=True):
     '''Return the names, versions and paths of all dependencies.
 
-    .. versionadded:: 1.7.0
+    .. versionadded:: (cfdm) 1.7.0
 
     :Parameters:
 
@@ -88,7 +88,7 @@ def CF():
     by this release of the cfdm.core package, and therefore the
     version can not be changed.
 
-    .. versionadded:: 1.7.0
+    .. versionadded:: (cfdm) 1.7.0
 
     :Returns:
 
