+++ cfdm/core/fieldancillary.py
@@ -23,14 +23,14 @@ class FieldAncillary(mixin.PropertiesData):
     definition, unlike, for instance, an auxiliary coordinate
     construct or domain ancillary construct.
 
-    .. versionadded:: 1.7.0
+    .. versionadded:: (cfdm) 1.7.0
 
     '''
     @property
     def construct_type(self):
         '''Return a description of the construct type.
 
-    .. versionadded:: 1.7.0
+    .. versionadded:: (cfdm) 1.7.0
 
     :Returns:
 
