### 2020-08-07 8a663b1
Reduce duplication of netCDF formats in netcdfwrite
