### 2020-08-07 4673604
Enable iterations over the core logic of netcdfwrite.write()
