### 2020-08-07 e7e2a65
Add keys to write_vars to indicate a (post) dry run iteration
