+++ cfdm/read_write/netcdf/netcdfwrite.py
@@ -41,7 +41,7 @@ class NetCDFWrite(IOWrite):
     def cf_geometry_types(self):
         '''Geometry types
 
-    .. versionadded:: 1.8.0
+    .. versionadded:: (cfdm) 1.8.0
 
         '''
         return set((
@@ -65,7 +65,7 @@ class NetCDFWrite(IOWrite):
     def _create_netcdf_group(self, nc, group_name):
         '''TODO
 
-    .. versionadded:: 1.8.6
+    .. versionadded:: (cfdm) 1.8.6
 
     :Parameters:
 
@@ -86,7 +86,7 @@ class NetCDFWrite(IOWrite):
         #                            force_use_existing=False):
         '''TODO
 
-    .. versionadded:: 1.7.0
+    .. versionadded:: (cfdm) 1.7.0
 
     :Parameters:
 
@@ -125,7 +125,7 @@ class NetCDFWrite(IOWrite):
     def _netcdf_name(self, base, dimsize=None, role=None):
         '''Return a new netCDF variable or dimension name.
 
-    .. versionadded:: 1.7.0
+    .. versionadded:: (cfdm) 1.7.0
 
     :Parameters:
 
@@ -191,7 +191,7 @@ class NetCDFWrite(IOWrite):
     def _numpy_compressed(self, array):
         '''Return all the non-masked data as a 1-d array.
 
-    .. versionadded:: 1.8.0
+    .. versionadded:: (cfdm) 1.8.0
 
     :Parameters:
 
@@ -900,7 +900,7 @@ class NetCDFWrite(IOWrite):
     def _create_geometry_container(self, field):
         '''TODO
 
-    .. versionadded:: 1.8.0
+    .. versionadded:: (cfdm) 1.8.0
 
     :Parameters:
 
@@ -1072,7 +1072,7 @@ class NetCDFWrite(IOWrite):
     When `True` is returned, the input variable is added to the
     g['seen'] dictionary.
 
-    .. versionadded:: 1.7.0
+    .. versionadded:: (cfdm) 1.7.0
 
     :Parameters:
 
@@ -1116,7 +1116,7 @@ class NetCDFWrite(IOWrite):
     def _write_geometry_container(self, field, geometry_container):
         '''Write a netCDF geometry container variable.
 
-    .. versionadded:: 1.8.0
+    .. versionadded:: (cfdm) 1.8.0
 
     :Returns:
 
@@ -1165,7 +1165,7 @@ class NetCDFWrite(IOWrite):
     dimension if required. Return the bounds variable's netCDF
     variable name.
 
-    .. versionadded:: 1.7.0
+    .. versionadded:: (cfdm) 1.7.0
 
     :Parameters:
 
@@ -1351,7 +1351,7 @@ class NetCDFWrite(IOWrite):
     * A netCDF part node count variable, if required.
     * A netCDF interior ring variable, if required.
 
-    .. versionadded:: 1.8.0
+    .. versionadded:: (cfdm) 1.8.0
 
     :Parameters:
 
@@ -1519,7 +1519,7 @@ class NetCDFWrite(IOWrite):
                           encodings):
         '''Create a netCDF node count variable.
 
-    .. versionadded:: 1.8.0
+    .. versionadded:: (cfdm) 1.8.0
 
     :Parameters:
 
@@ -1604,7 +1604,7 @@ class NetCDFWrite(IOWrite):
         '''Get the base of the netCDF dimension for part node count and
     interior ring variables.
 
-    .. versionadded:: 1.8.0
+    .. versionadded:: (cfdm) 1.8.0
 
     :Returns:
 
@@ -1644,7 +1644,7 @@ class NetCDFWrite(IOWrite):
[…]
