+++ setup.py
@@ -137,6 +137,7 @@ setup(name = "cfdm",
                       'cfdm.core.abstract',
                       'cfdm.core.data',
                       'cfdm.core.data.abstract',
+                      'cfdm.core.meta',
                       'cfdm.core.mixin',
                       'cfdm.data',
                       'cfdm.data.abstract',
