### 2020-08-10 ffc0946
docstring inheritance
