+++ cfdm/cellmeasure.py
@@ -193,30 +193,7 @@ class CellMeasure(mixin.NetCDFVariable,
             If True then the ``_FillValue`` and ``missing_value``
             properties are omitted from the comparison.
 
-        verbose: `int` or `str` or `None`, optional
-            If an integer from ``-1`` to ``3``, or an equivalent string
-            equal ignoring case to one of:
-
-            * ``'DISABLE'`` (``0``)
-            * ``'WARNING'`` (``1``)
-            * ``'INFO'`` (``2``)
-            * ``'DETAIL'`` (``3``)
-            * ``'DEBUG'`` (``-1``)
-
-            set for the duration of the method call only as the minimum
-            cut-off for the verboseness level of displayed output (log)
-            messages, regardless of the globally-configured `cfdm.log_level`.
-            Note that increasing numerical value corresponds to increasing
-            verbosity, with the exception of ``-1`` as a special case of
-            maximal and extreme verbosity.
-
-            Otherwise, if `None` (the default value), output messages will
-            be shown according to the value of the `cfdm.log_level` setting.
-
-            Overall, the higher a non-negative integer or equivalent string
-            that is set (up to a maximum of ``3``/``'DETAIL'``) for
-            increasing verbosity, the more description that is printed to
-            convey information about differences that lead to inequality.
+        {{verbose: `int` or `str` or `None`, optional}}
 
         ignore_properties: sequence of `str`, optional
             The names of properties to omit from the comparison.
@@ -233,12 +210,7 @@ class CellMeasure(mixin.NetCDFVariable,
             the arrays in their uncompressed forms. By default only
             the arrays in their uncompressed forms are compared.
 
-        ignore_type: `bool`, optional
-            Any type of object may be tested but, in general, equality is
-            only possible with another cell measure construct, or a
-            subclass of one. If *ignore_type* is True then
-            ``CellMeasure(source=other)`` is tested, rather than the
-            ``other`` defined by the *other* parameter.
+        {{ignore_type: `bool`, optional}}
 
     :Returns:
 
