+++ cfdm/core/field.py
@@ -73,7 +73,7 @@ class Field(mixin.ConstructAccess,
             *source* parameter is set.
 
             Properties may also be set after initialisation with the
-            `set_properties` and `set_property` methods.
+            `set_properties` and `set_property` methods.""",
 
             *Parameter example:*
                ``properties={'standard_name': 'air_temperature'}``
