+++ cfdm/core/abstract/propertiesdatabounds.py
@@ -2,7 +2,7 @@ import abc
 
 import numpy
 
-from . import PropertiesData
+from .propertiesdata import PropertiesData
 
 
 # --------------------------------------------------------------------
@@ -31,7 +31,7 @@ class PropertiesDataBounds(PropertiesData):
             *source* parameter is set.
 
             Properties may also be set after initialisation with the
-            `set_properties` and `set_property` methods.
+            `set_properties` and `set_property` methods.""",
 
             *Parameter example:*
                ``properties={'standard_name': 'longitude'}``
