### 2020-08-11 d94eccd
docstring subs in separate files
