+++ cfdm/test/test_docstring.py
@@ -76,7 +76,7 @@ class DocstringTest(unittest.TestCase):
         )
 
     def test_docstring(self):
-        # Test that all {{ and }} occurences have been substituted
+        # Test that all {{ occurences have been substituted
         for klass in self.subclasses_of_Container:
             for x in (klass, klass()):
                 for name in dir(x):
@@ -94,11 +94,6 @@ class DocstringTest(unittest.TestCase):
 
                     self.assertNotIn(
                         '{{', f.__doc__,
-                        '\n\nCLASS: {}\nMETHOD NAME: {}\nMETHOD: {}'.format(
-                            klass, name, f))
-
-                    self.assertNotIn(
-                        '}}', f.__doc__,
                         '\n\nCLASS: {}\nMETHOD NAME: {}\nMETHOD: {}'.format(
                             klass, name, f))
 
@@ -168,10 +163,10 @@ class DocstringTest(unittest.TestCase):
                     (1, 2)
                 )
 
-    def test_docstring_docstring_substitutions(self):
+    def test_docstring_docstring_substitution(self):
         for klass in self.subclasses_of_Container:
             for x in (klass, klass()):
-                d = x._docstring_substitutions()
+                d = x._docstring_substitution()
                 self.assertIsInstance(d, dict)
                 self.assertIn('{{repr}}', d)
 
