+++ cfdm/core/docstring/__init__.py
@@ -1,0 +1,2 @@
+from .docstring import _docstring_substitution_definitions
+
