### 2020-08-11 8555aea
docstring inheritance
