+++ cfdm/core/docstring/docstring.py
@@ -6,7 +6,7 @@ with the replacement text defined by the corresponding value.
 Keys must be `str` or `re.Pattern` objects.
 
 If a key is a `str` then the corresponding value must be a string.
-    
+
 If a key is a `re.Pattern` object then the corresponding value must be
 a string or a callable, as accepted by the `re.Pattern.sub` method.
 
@@ -20,9 +20,9 @@ replacement text, and will be substituted as ususal.
 _docstring_substitution_definitions = {
     '{{repr}}':
     '',
-    
+
     '{{default: optional}}':
-            '''default: optional
+    '''default: optional
             Return the value of the *default* parameter if data have
             not been set. If set to an `Exception` instance then it
             will be raised instead.''',
