+++ cfdm/core/docstring/__init__.py
@@ -1,2 +1,1 @@
 from .docstring import _docstring_substitution_definitions
-
