### 2020-08-11 be4ccc9
docstring inheritance
