+++ cfdm/mixin/container.py
@@ -71,7 +71,7 @@ class Container:
     :Returns:
 
         `dict`
-            The doctring substitutions that have been applied.
+            The docstring substitutions that have been applied.
 
         '''
         return _docstring_substitution_definitions
