+++ cfdm/constructs.py
@@ -3,6 +3,7 @@ import textwrap
 from copy import deepcopy
 
 from . import core
+from . import mixin
 
 from .decorators import _manage_log_level_via_verbosity
 
@@ -10,7 +11,8 @@ from .decorators import _manage_log_level_via_verbosity
 logger = logging.getLogger(__name__)
 
 
-class Constructs(core.Constructs):
+class Constructs(mixin.Container,
+                 core.Constructs):
     '''A container for metadata constructs.
 
     Calling a `Constructs` instance selects metadata constructs by
@@ -49,10 +51,10 @@ class Constructs(core.Constructs):
 
     >>> print(c('latitude'))
     Constructs:
-    {'dimensioncoordinate0': <DimensionCoordinate: latitude(5) degrees_north>}
+    {'dimensioncoordinate0': <{{repr}}DimensionCoordinate: latitude(5) degrees_north>}
     >>> print(c.filter_by_identity('latitude'))
     Constructs:
-    {'dimensioncoordinate0': <DimensionCoordinate: latitude(5) degrees_north>}
+    {'dimensioncoordinate0': <{{repr}}DimensionCoordinate: latitude(5) degrees_north>}
 
     See `filter_by_identity` for more examples.
 
@@ -414,6 +416,7 @@ class Constructs(core.Constructs):
 
     :Returns:
 
+        `{{class}}`
             The deep copy.
 
     **Examples:**
@@ -535,22 +538,14 @@ class Constructs(core.Constructs):
                _return_axis_map=False):
         '''Whether two `Constructs` instances are the same.
 
-    Two real numbers ``x`` and ``y`` are considered equal if
-    ``|x-y|<=atol+rtol|y|``, where ``atol`` (the tolerance on absolute
-    differences) and ``rtol`` (the tolerance on relative differences)
-    are positive, typically very small numbers. See the *atol* and
-    *rtol* parameters.
+    {{equals tolerance}}
 
-    Any compression is ignored by default, with only the arrays in
-    their uncompressed forms being compared. See the
-    *ignore_compression* parameter.
+    {{equals compression}}
 
     Any type of object may be tested but equality is only possible
     with another `Constructs` construct, or a subclass of one.
 
-    NetCDF elements, such as netCDF variable and dimension names, do
-    not constitute part of the CF data model and so are not checked on
-    any construct.
+    {{equals netCDF}}
 
     .. versionadded:: (cfdm) 1.7.0
 
@@ -576,17 +571,9 @@ class Constructs(core.Constructs):
 
         {{verbose: `int` or `str` or `None`, optional}}
 
-        ignore_data_type: `bool`, optional
-            If True then ignore the data types in all numerical
-            comparisons. By default different numerical data types
-            imply inequality, regardless of whether the elements are
-            within the tolerance for equality.
+        {{ignore_data_type: `bool`, optional}}
 
-        ignore_compression: `bool`, optional
-            If False then the compression type and, if applicable, the
-            underlying compressed arrays must be the same, as well as
-            the arrays in their uncompressed forms. By default only
[…]
