### 2020-08-12 d5161dc
added docstring subs
