+++ cfdm/mixin/propertiesdatabounds.py
@@ -149,6 +149,7 @@ class PropertiesDataBounds(PropertiesData):
 
     :Returns:
 
+        `{{class}}`
             The subspace of the construct.
 
     **Examples:**
@@ -480,11 +481,11 @@ class PropertiesDataBounds(PropertiesData):
 
     :Parameters:
 
-        inplace: `bool`, optional
-            If True then do the operation in-place and return `None`.
+        {{inplace: `bool`, optional}}
 
     :Returns:
 
+        `{{class}}` or `None`
             A new instance with masked values, or `None` if the
             operation was in-place.
 
@@ -541,25 +542,23 @@ class PropertiesDataBounds(PropertiesData):
 
     :Parameters:
 
-        default: optional
-            Return the value of the *default* parameter if the node
-            count variable has not been set. If set to an `Exception`
-            instance then it will be raised instead.
+        {{default: optional}}
 
     :Returns:
 
+        `NodeCount`
             The removed node count variable.
 
     **Examples:**
 
-    >>> n = cfdm.NodeCount(properties={'long_name': 'node counts'})
+    >>> n = {{package}}.NodeCount(properties={'long_name': 'node counts'})
     >>> c.set_node_count(n)
     >>> c.has_node_count()
     True
     >>> c.get_node_count()
-    <NodeCount: long_name=node counts>
+    <{{repr}}NodeCount: long_name=node counts>
     >>> c.del_node_count()
-    <NodeCount: long_name=node counts>
+    <{{repr}}NodeCount: long_name=node counts>
     >>> c.has_node_count()
     False
 
@@ -582,25 +581,23 @@ class PropertiesDataBounds(PropertiesData):
 
     :Parameters:
 
-        default: optional
-            Return the value of the *default* parameter if the part
-            node count variable has not been set. If set to an
-            `Exception` instance then it will be raised instead.
+        {{default: optional}}
 
     :Returns:
 
-            The removed part_node count variable.
+        `PartNodeCount`
+            The removed part node count variable.
 
     **Examples:**
 
-    >>> p = cfdm.PartNodeCount(properties={'long_name': 'part node counts'})
+    >>> p = {{package}}.PartNodeCount(properties={'long_name': 'part node counts'})
     >>> c.set_part_node_count(p)
     >>> c.has_part_node_count()
     True
     >>> c.get_part_node_count()
-    <PartNodeCount: long_name=part node counts>
+    <{{repr}}PartNodeCount: long_name=part node counts>
     >>> c.del_part_node_count()
-    <PartNodeCount: long_name=part node counts>
+    <{{repr}}PartNodeCount: long_name=part node counts>
     >>> c.has_part_node_count()
     False
 
@@ -626,9 +623,7 @@ class PropertiesDataBounds(PropertiesData):
 
     :Returns:
 
-            The description. If *display* is True then the description
-            is printed and `None` is returned. Otherwise the
-            description is returned as a string.
+        {{returns dump}}
 
         '''
         # ------------------------------------------------------------
@@ -869,25 +864,23 @@ class PropertiesDataBounds(PropertiesData):
 
     :Parameters:
 
-        default: optional
-            Return the value of the *default* parameter if a node
-            count variable has not been set. If set to an `Exception`
-            instance then it will be raised instead.
+        {{default: optional}}
 
     :Returns:
 
+        `NodeCount`
             The node count variable.
 
     **Examples:**
 
[…]
