+++ cfdm/mixin/container.py
@@ -1,4 +1,3 @@
-import inspect
 import logging
 
 import numpy
