+++ cfdm/test/test_List.py
@@ -1,5 +1,4 @@
 import datetime
-import os
 import unittest
 
 import cfdm
