+++ cfdm/test/test_core_1.py
@@ -1,6 +1,4 @@
 import datetime
-import os
-import sys
 import unittest
 
 import numpy
