### 2020-08-18 fe1628a
interior ring docs
