### 2020-08-18 f1067af
docs string substitutions
