+++ cfdm/test/test_dsg.py
@@ -301,7 +301,7 @@ class DSGTest(unittest.TestCase):
 
 
 if __name__ == '__main__':
-    print('Run date:', datetime.datetime.utcnow())
+    print('Run date:', datetime.datetime.now())
     cfdm.environment()
     print()
     unittest.main(verbosity=2)
