+++ cfdm/test/setup_create_field_2.py
@@ -250,6 +250,6 @@ class create_fieldTest_2(unittest.TestCase):
 
 if __name__ == "__main__":
     print('Run date:', datetime.datetime.now())
-    cfdm.environment(display=False)
+    cfdm.environment()
     print('')
     unittest.main(verbosity=2)
