+++ cfdm/fieldancillary.py
@@ -46,28 +46,17 @@ class FieldAncillary(mixin.NetCDFVariable,
 
     :Parameters:
 
-        properties: `dict`, optional
-            Set descriptive properties. The dictionary keys are
-            property names, with corresponding values. Ignored if the
-            *source* parameter is set.
-
-            Properties may also be set after initialisation with the
-            `set_properties` and `set_property` methods.""",
+        {{init properties: `dict`, optional}}
 
            *Parameter example:*
               ``properties={'standard_name': 'altitude'}``
 
-        data: `Data`, optional
-            Set the data array. Ignored if the *source* parameter is
-            set.
-
-            The data array may also be set after initialisation with
-            the `set_data` method.
+        {{init data: `{{+Data}}`, optional}}
 
         source: optional
             Initialize the properties and data from those of *source*.
 
-        copy: `bool`, optional
+        {{init copy: `bool`, optional}}
             If False then do not deep copy input parameters prior to
             initialization. By default arguments are deep copied.
 
