+++ cfdm/test/test_gathering.py
@@ -227,7 +227,7 @@ class GatheredTest(unittest.TestCase):
 
 
 if __name__ == '__main__':
-    print('Run date:', datetime.datetime.utcnow())
+    print('Run date:', datetime.datetime.now())
     cfdm.environment()
     print()
     unittest.main(verbosity=2)
