+++ cfdm/datum.py
@@ -54,9 +54,7 @@ class Datum(mixin.Parameters,
         source: optional
             Initialize the parameters from those of *source*.
 
-        copy: `bool`, optional
-            If False then do not deep copy input parameters prior to
-            initialization. By default arguments are deep copied.
+        {{init copy: `bool`, optional}}
 
         '''
         super().__init__(parameters=parameters,
