+++ cfdm/test/test_NodeCountProperties.py
@@ -40,7 +40,7 @@ class NodeCountPropertiesTest(unittest.TestCase):
 
 
 if __name__ == '__main__':
-    print('Run date:', datetime.datetime.utcnow())
+    print('Run date:', datetime.datetime.now())
     cfdm.environment()
     print()
     unittest.main(verbosity=2)
