+++ cfdm/abstract/__init__.py
@@ -1,1 +1,2 @@
+from .container import Container
 from .implementation import Implementation
