+++ cfdm/mixin/container.py
@@ -19,6 +19,9 @@ class Container:
     .. versionadded:: (cfdm) 1.7.0
 
     '''
+#    # Set the package depth for {{package}} docstring substitutions
+#    _docstring_package_depth = 0
+
     def __repr__(self):
         '''Called by the `repr` built-in function.
 
@@ -40,33 +43,16 @@ class Container:
         out = sorted(self._components)
         return ', '.join(out)
 
-    def __docstring_substitution__(self):
+    def __docstring_substitutions__(self):
         '''Define docstring substitutions that apply to this class and all of
     its subclasses.
 
     These are in addtion to, and take precendence over, docstring
     substitutions defined by the base classes of this class.
 
-    Text to be replaced is specified as a key in the returned
-    dictionary, with the replacement text defined by the corresponding
-    value.
-
-    Keys must be `str` or `re.Pattern` objects.
-
-    If a key is a `str` then the corresponding value must be a string.
-
-    If a key is a `re.Pattern` object then the corresponding value
-    must be a string or a callable, as accepted by the
-    `re.Pattern.sub` method.
-
-    Special docstring subtitutions, as defined by
-    `_special_docstring_substitutions`, may be used in the replacement
-    text, and will be substituted as ususal.
+    See `_docstring_substitutions` for details.
 
     .. versionaddedd:: (cfdm) 1.8.7.0
-
-    .. seealso:: `_docstring_substitutions`,
-                 `_special_docstring_substitutions`
 
     :Returns:
 
@@ -75,6 +61,14 @@ class Container:
 
         '''
         return _docstring_substitution_definitions
+
+    def __docstring_package_depth__(self):
+        '''Return the package depth for {{package}} docstring substitutions.
+
+    See `_docstring_package_depth` for details.
+
+        '''
+        return 0
 
     @property
     def _atol(self):
