+++ cfdm/index.py
@@ -63,7 +63,7 @@ class Index(mixin.NetCDFVariable,
             *Parameter example:*
               ``properties={'long_name': 'which station this obs is for'}``
 
-        {{init data: `{{+Data}}`, optional}}
+        {{init data: `Data`, optional}}
 
         source: optional
             Initialize the properties and data from those of *source*.
