+++ cfdm/core/meta/rewritedocstring.py
@@ -2,91 +2,6 @@ import inspect
 import re
 
 
-# --------------------------------------------------------------------
-# Define the "plus class" regular expression that matches, e.g.
-# {{+Data}}, with Data as the grouped match.
-# --------------------------------------------------------------------
-_plus_class_regex = re.compile('{{\+(\w.*?)}}')
-
-
-def _docstring_replacement_class(match):
-    '''Return the first of the match groups.
-
-    .. versionadded:: (cfdm) 1.8.7.0
-
-        '''
-    return match.group(1)
-
-
-def _docstring_replacement_core_class(match):
-    '''Return the first of the match groups prefixed by 'core.'
-
-    .. versionadded:: (cfdm) 1.8.7.0
-
-        '''
-    return 'core.' + match.group(1)
-
-
-def _docstring_update(class_name, f, method_name, module, config):
-    '''TODO
-
-    .. versionadded:: (cfdm) 1.8.7.0
-
-    '''
-    doc = f.__doc__
-    if doc is None:
-        return
-
-    # ------------------------------------------------------------
-    # Find the package name, class name, and whether or we are in
-    # the cfdm.core package.
-    # ------------------------------------------------------------
-    core = False
-    module = module.split('.')
-    if len(module) >= 2:
-        package_name, package2 = module[:2]
-        if package_name == 'cfdm' and package2 == 'core':
-            class_name = 'core.' + class_name
-            core = True
-    else:
-        package_name = module[0]
-
-    # ------------------------------------------------------------
-    # Do general substitutions first
-    # ------------------------------------------------------------
-    for key, value in config.items():
-        try:
-            # Compiled regular expression substitution
-            doc = key.sub(value, doc)
-        except AttributeError:
-            # String substitution
-            doc = doc.replace(key, value)
-    # --- End: for
-
-    # ------------------------------------------------------------
-    # Now do special substitutions
-    # ------------------------------------------------------------
-    # Insert the name of the package
-    doc = doc.replace('{{package}}', package_name)
-
-    # Insert the name of the class containing this method
-    doc = doc.replace('{{class}}', class_name)
-
-    # Insert the name of a different class
-    if '{{+' in doc:
-        if core:
-            func = _docstring_replacement_core_class
-        else:
-            func = _docstring_replacement_class
-
-        doc = _plus_class_regex.sub(func, doc)
-
-    # ----------------------------------------------------------------
-    # Set the rewritten docstring on the method
-    # ----------------------------------------------------------------
-    f.__doc__ = doc
-
-
 class RewriteDocstringMeta(type):
     '''Modify docstrings.
 
@@ -117,41 +32,84 @@ class RewriteDocstringMeta(type):
         '''
         # ------------------------------------------------------------
         # Combine the docstring substitutions from all classes in the
-        # inheritance tree.
-        #
-        # The value for a key that occurs in multiple classes will be
-        # taken from the class closest to the child class.
+        # inheritance tree. The value for a key that occurs in
+        # multiple classes will be taken from the class closest to the
+        # child class.
         # ------------------------------------------------------------
         docstring_rewrite = {}
+
         for parent in parents[::-1]:
             parent_docstring_rewrite = getattr(
-                parent, '_docstring_substitution', None)
+                parent, '_docstring_substitutions', None)
             if parent_docstring_rewrite is not None:
-                docstring_rewrite.update(parent_docstring_rewrite())
+                docstring_rewrite.update(parent_docstring_rewrite(parent))
             else:
                 parent_docstring_rewrite = getattr(
-                    parent, '__docstring_substitution__', None)
+                    parent, '__docstring_substitutions__', None)
                 if parent_docstring_rewrite is not None:
[…]
