+++ cfdm/core/abstract/propertiesdata.py
@@ -20,7 +20,7 @@ class PropertiesData(Properties):
             *Parameter example:*
               ``properties={'standard_name': 'altitude'}``
 
-        {{init data: `{{+Data}}`, optional}}
+        {{init data: `Data`, optional}}
 
         source: optional
             Initialize the properties and data from those of *source*.
@@ -53,7 +53,7 @@ class PropertiesData(Properties):
 
     ``f.data`` is equivalent to ``f.get_data()``
 
-    Note that a `{{+Data}}` instance is returned. Use its `array`
+    Note that a `Data` instance is returned. Use its `array`
     attribute to return the data as a `numpy` array.
 
     The units, calendar and fill value properties are, if set,
@@ -61,24 +61,24 @@ class PropertiesData(Properties):
 
     .. versionadded:: (cfdm) 1.7.0
 
-    .. seealso:: `Data.array`, `del_data`, `get_data`, `has_data`,
-                 `set_data`
+    .. seealso:: `Data.array`, `del_data`, `get_data`,
+                 `has_data`, `set_data`
 
     :Returns:
 
-        `{{+Data}}`
+        `Data`
             The data.
 
     **Examples:**
 
     >>> import numpy
     >>> f = {{package}}.{{class}}()
-    >>> f.set_data({{package}}.{{+Data}}(numpy.arange(9.)))
+    >>> f.set_data({{package}}.Data(numpy.arange(9.)))
     >>> f.has_data()
     True
     >>> d = f.data
     >>> d
-    <{{repr}}{{+Data}}(10): [0.0, ..., 9.0]>
+    <{{repr}}Data(10): [0.0, ..., 9.0]>
     >>> f.data.shape
     (10,)
 
@@ -93,7 +93,7 @@ class PropertiesData(Properties):
 
     ``f.copy()`` is equivalent to ``copy.deepcopy(f)``.
 
-    Arrays within `{{+Data}}` instances are copied with a
+    Arrays within `Data` instances are copied with a
     copy-on-write technique. This means that a copy takes up very
     little extra memory, even when the original contains very large
     data arrays, and the copy operation is fast.
@@ -134,20 +134,20 @@ class PropertiesData(Properties):
 
     :Returns:
 
-        `{{+Data}}`
+        `Data`
             The removed data.
 
     **Examples:**
 
     >>> f = {{package}}.{{class}}()
-    >>> d = {{package}}.{{+Data}}(range(10))
+    >>> d = {{package}}.Data(range(10))
     >>> f.set_data(d)
     >>> f.has_data()
     True
     >>> f.get_data()
-    <{{repr}}{{+Data}}(10): [0, ..., 9]>
+    <{{repr}}Data(10): [0, ..., 9]>
     >>> f.del_data()
-    <{{repr}}{{+Data}}(10): [0, ..., 9]>
+    <{{repr}}Data(10): [0, ..., 9]>
     >>> f.has_data()
     False
     >>> print(f.get_data(None))
@@ -164,7 +164,7 @@ class PropertiesData(Properties):
                  _fill_value=True):
         '''Return the data.
 
-    Note that a `{{+Data}}` instance is returned. Use its `array`
+    Note that a `Data` instance is returned. Use its `array`
     attribute to return the data as an independent `numpy` array.
 
     The units, calendar and fill value properties are, if set,
@@ -172,7 +172,7 @@ class PropertiesData(Properties):
 
     .. versionadded:: (cfdm) 1.7.0
 
[…]

[…]
