+++ cfdm/data/mixin/arraymixin.py
@@ -7,9 +7,6 @@ class ArrayMixin:
     .. versionadded:: (cfdm) 1.8.7.0
 
     '''
-#    # Set the package depth for {{package}} docstring substitutions
-#    _docstring_package_depth = 0
-
     def __array__(self, *dtype):
         '''The numpy array interface.
 
@@ -83,7 +80,11 @@ class ArrayMixin:
         return "shape={0}, dtype={1}".format(self.shape, self.dtype)
 
     def __docstring_package_depth__(self):
-        '''TODO'''
+        '''Return the package depth for {{package}} docstring substitutions.
+
+    See `_docstring_package_depth` for details.
+
+        '''
         return 0
 
     def get_compression_type(self):
