### 2020-08-19 df84c16
Consistent functionality descriptions, incl. trailing ', and' removal
