### 2020-08-26 4fe5e9c
Docs: fix all cases of missing syntax highl. due to extra spaces
Due to '>>>  <code>' rather than '>>> <code>', fixed with:
  $ find . -type f | xargs sed -i 's/>>>  />>> /g'
