+++ cfdm/mixin/propertiesdata.py
@@ -164,7 +164,7 @@ class PropertiesData(Properties):
     >>> f.size
     7008
 
-    >>>  f.shape
+    >>> f.shape
     (73,)
     >>> f.ndim
     1
