+++ cfdm/data/data.py
@@ -1002,7 +1002,7 @@ class Data(mixin.Container,
      [4 --  6  7]
      [8  9 10 11]]
 
-    >>>  print(d.apply_masking().array)
+    >>> print(d.apply_masking().array)
     [[0  1  2  3]
      [4 --  6  7]
      [8  9 10 11]]
