### 2020-08-28 d832d5d
API ref: add missing Data.{max, min} aliases to class listing
