### 2020-09-01 c6e2bee
core set_data inplace=True
