+++ docs/1.8.6.0/_sources/method/cfdm.core.CellMeasure.get_measure.rst.txt
@@ -1,7 +1,0 @@
-cfdm.core.CellMeasure.get_measure
-=================================
-
-.. currentmodule:: cfdm
-.. default-role:: obj
-
-.. automethod:: cfdm.core.CellMeasure.get_measure
