+++ docs/1.8.6.0/_sources/method/cfdm.Field.nc_del_component_dimension.rst.txt
@@ -1,7 +1,0 @@
-cfdm.Field.nc_del_component_dimension
-=====================================
-
-.. currentmodule:: cfdm
-.. default-role:: obj
-
-.. automethod:: cfdm.Field.nc_del_component_dimension
