+++ docs/1.8.6.0/_sources/method/cfdm.core.abstract.PropertiesData.set_data.rst.txt
@@ -1,7 +1,0 @@
-cfdm.core.abstract.PropertiesData.set_data
-==========================================
-
-.. currentmodule:: cfdm
-.. default-role:: obj
-
-.. automethod:: cfdm.core.abstract.PropertiesData.set_data
