+++ docs/1.8.6.0/_sources/method/cfdm.core.CellMeasure.set_properties.rst.txt
@@ -1,7 +1,0 @@
-cfdm.core.CellMeasure.set_properties
-====================================
-
-.. currentmodule:: cfdm
-.. default-role:: obj
-
-.. automethod:: cfdm.core.CellMeasure.set_properties
