+++ docs/_sources/method/cfdm.core.Datum.set_parameter.rst.txt
@@ -1,7 +1,0 @@
-cfdm.core.Datum.set_parameter
-=============================
-
-.. currentmodule:: cfdm
-.. default-role:: obj
-
-.. automethod:: cfdm.core.Datum.set_parameter
