+++ docs/1.8.6.0/_sources/attribute/cfdm.RaggedIndexedArray.compressed_array.rst.txt
@@ -1,7 +1,0 @@
-cfdm.RaggedIndexedArray.compressed_array
-========================================
-
-.. currentmodule:: cfdm
-.. default-role:: obj
-
-.. autoattribute:: cfdm.RaggedIndexedArray.compressed_array
