+++ docs/1.8.6.0/_sources/attribute/cfdm.Field.dtype.rst.txt
@@ -1,7 +1,0 @@
-cfdm.Field.dtype
-================
-
-.. currentmodule:: cfdm
-.. default-role:: obj
-
-.. autoattribute:: cfdm.Field.dtype
