+++ docs/1.8.6.0/_sources/method/cfdm.Bounds.nc_has_dimension.rst.txt
@@ -1,7 +1,0 @@
-cfdm.Bounds.nc_has_dimension
-============================
-
-.. currentmodule:: cfdm
-.. default-role:: obj
-
-.. automethod:: cfdm.Bounds.nc_has_dimension
