+++ docs/_sources/class/cfdm.Count.rst.txt
@@ -1,155 +1,0 @@
-.. currentmodule:: cfdm
-.. default-role:: obj
-
-cfdm.Count
-==========
-
-----
-
-.. autoclass:: cfdm.Count
-   :no-members:
-   :no-inherited-members:
-
-Inspection
-----------
-
-.. rubric:: Methods
-	    
-.. autosummary::
-   :nosignatures:
-   :toctree: ../method/
-   :template: method.rst
-
-   ~cfdm.Count.dump
-   ~cfdm.Count.identity  
-   ~cfdm.Count.identities
-  
-Properties
-----------
-
-.. rubric:: Methods
-	    
-.. autosummary::
-   :nosignatures:
-   :toctree: ../method/
-   :template: method.rst
-
-   ~cfdm.Count.del_property
-   ~cfdm.Count.get_property
-   ~cfdm.Count.has_property
-   ~cfdm.Count.set_property
-   ~cfdm.Count.properties
-   ~cfdm.Count.clear_properties
-   ~cfdm.Count.set_properties
-
-Data
-----
-
-.. rubric:: Methods
-	    
-.. autosummary::
-   :nosignatures:
-   :toctree: ../method/
-   :template: method.rst
-
-   ~cfdm.Count.del_data
-   ~cfdm.Count.get_data
-   ~cfdm.Count.has_data
-   ~cfdm.Count.set_data  
-   ~cfdm.Count.insert_dimension
-   ~cfdm.Count.squeeze
-   ~cfdm.Count.transpose
-   
-.. rubric:: Attributes
-   
-.. autosummary::
-   :nosignatures:
-   :toctree: ../attribute/
-   :template: attribute.rst
-
-   ~cfdm.Count.data
-
-Miscellaneous
--------------
-
-.. rubric:: Methods
-	    
-.. autosummary::
-   :nosignatures:
-   :toctree: ../method/
-   :template: method.rst
-
-   ~cfdm.Count.copy
-   ~cfdm.Count.equals
-   ~cfdm.Count.get_filenames
-
-NetCDF
-------
-
-.. rubric:: Methods
-	    
-.. autosummary::
-   :nosignatures:
-   :toctree: ../method/
-   :template: method.rst
-
-   ~cfdm.Count.nc_del_variable
-   ~cfdm.Count.nc_get_variable
-   ~cfdm.Count.nc_has_variable
-   ~cfdm.Count.nc_set_variable
-   ~cfdm.Count.nc_del_dimension
-   ~cfdm.Count.nc_get_dimension
-   ~cfdm.Count.nc_has_dimension
-   ~cfdm.Count.nc_set_dimension
-
-Groups
-^^^^^^
-
-.. rubric:: Methods
-	    
-.. autosummary::
-   :nosignatures:
-   :toctree: ../method/
-   :template: method.rst
-
-   ~cfdm.Count.nc_variable_groups
-   ~cfdm.Count.nc_clear_variable_groups
-   ~cfdm.Count.nc_set_variable_groups
-   ~cfdm.Count.nc_dimension_groups
-   ~cfdm.Count.nc_clear_dimension_groups
-   ~cfdm.Count.nc_set_dimension_groups
-   
-Sample dimension
-^^^^^^^^^^^^^^^^
-
-.. rubric:: Methods
-	    
-.. autosummary::
-   :nosignatures:
-   :toctree: ../method/
-   :template: method.rst
-
-   ~cfdm.Count.nc_del_sample_dimension
-   ~cfdm.Count.nc_get_sample_dimension
-   ~cfdm.Count.nc_has_sample_dimension
-   ~cfdm.Count.nc_set_sample_dimension
[…]
