+++ docs/1.8.6.0/_sources/method/cfdm.Index.nc_clear_variable_groups.rst.txt
@@ -1,7 +1,0 @@
-cfdm.Index.nc_clear_variable_groups
-===================================
-
-.. currentmodule:: cfdm
-.. default-role:: obj
-
-.. automethod:: cfdm.Index.nc_clear_variable_groups
