+++ cfdm/cellmeasure.py
@@ -67,7 +67,9 @@ class CellMeasure(mixin.NetCDFVariable,
            *Parameter example:*
              ``properties={'standard_name': 'cell_area'}``
 
-        {{init data: `Data`, optional}}
+        {{init data: data_like, optional}}
+
+            {{data_like}}
 
         source: optional
             Initialize the measure, properties and data from those of
