### 2020-09-02 d5192a0
Correct the stated default for 'copy' kwarg of some methods
