### 2020-09-03 ae8b4f7
Daily docstring TODOs: netcdfread parsing utilities
