### 2020-09-03 927d6c8
Daily docstring TODOs: netcdfread creation methods
