### 2020-09-04 337d230
docstring rewriter reorganisation and docs
