+++ docs/source/class.rst
@@ -95,4 +95,4 @@ Classes that support the creation and storage of compressed arrays.
    cfdm.RaggedIndexedArray
    cfdm.RaggedIndexedContiguousArray
    cfdm.CompressedArray
-   
+
