### 2020-09-06 #2531 @bradyrx
@usethedata, @davidhassell

Hi all, thanks for the patience here. It looks like this is stuck in pre-review? I have time now to do this so I'm just going to post my review here using the checklist I used last time I reviewed for JOSS.

Overall, `cfdm` looks great and I'm happy to suggest its acceptance to JOSS following some minor corrections I've listed below. Some are nitpicky grammar on the paper, others are some thoughts to make contributions more accessible to early open-source contributors. I was able to use `cfdm` very quickly with some forecasting data I had saved out to disk. The docs are thorough and give good tutorials. I didn't go too deep into the code base, but after skimming through, it hits the points I care about -- robust testing, leverages inheritance, and has consistently good docstrings.

Yes. The source code is available and the repo is very well structured for open-source science.

Yes, it looks like the two main contributors to the code base are the two authors.

Yes, I loaded in some SubX subseasonal forecasting output I had and was able to follow along with the tutorial and inspect its properties.

Yes, I tried playing around with inplace operations (in memory) per the docs and things were faster in place.

Really impressed by the thorough tutorial examples on the docs. Well done!

There's no `codecov` or `coveralls` indicator of how much coverage the testing gives, but I can see that there is a thorough amount of testing done under `cfdm/test`.

1. I think it would be good to add a code of conduct. See `Insights/Community'. That's the main thing missing (outside of PR templates). They have some stock ones that work well. Ours at `climpred` is based on https://www.contributor-covenant.org/version/1/4/code-of-conduct/.
2. It looks like `CONTRIBUTING.md` in the main repo is decoupled from the "support" page on the docs. I.e., the docs are missing some of this information. I would also suggest adding more information for first-time or early contributors, since there's a lot of steps that go into contributing to a python package on Github. I have a lightweight personal package with some contribution instructions that you're welcome to pull from: https://esmtools.readthedocs.io/en/stable/contributing.html (see "Preparing Pull Requests").
3. I might change the docs page name from "Support" to something more obvious like "How to Contribute".

It is mentioned that `cfdm` is the only library that is guaranteed to handle every possible type of CF-compliant dataset. I think this leaves out popular high-level packages like `iris` (https://github.com/SciTools/iris) that have CF-compliant manipulation of Earth System data, and it doesn't look like `cfdm` is a dependency there (please correct me if I'm wrong). `xarray` doesn't natively handle CF well, but an extension to it `cf-xarray` does: https://github.com/xarray-contrib/cf-xarray. So there might be some other packages that can be nodded to that think of the CF data model.

Generally well written. See my bottom notes on small fixes to the paper's grammar. There are some spurious semicolons and commas in a few places throughout. Please double check uses of `,` and `;`.

Some notes on the paper:
* Under the summary, the semicolon should be dropped from "They cater for data from model simulations; and from observations, ...". Maybe change the "and" following it to "as well as" for flow.
* The sentence following the previous is structured somewhat awkwardly. I would say something like "CF conventions provide a description of the physical meaning of netCDF data variables, as well as their spatial, temporal, and other dimensional properties."
