+++ cfdm/mixin/propertiesdatabounds.py
@@ -431,7 +431,7 @@ class PropertiesDataBounds(PropertiesData):
     # ----------------------------------------------------------------
     # Methods
     # ----------------------------------------------------------------
-    @_inplace_enabled
+    @_inplace_enabled(default=False)
     def apply_masking(self, bounds=True, inplace=False):
         '''Apply masking as defined by the CF conventions.
 
@@ -1180,7 +1180,7 @@ class PropertiesDataBounds(PropertiesData):
 
         return bounds.get_data(default=default)
 
-    @_inplace_enabled
+    @_inplace_enabled(default=False)
     def insert_dimension(self, position, inplace=False):
         '''Expand the shape of the data array.
 
@@ -1337,7 +1337,7 @@ class PropertiesDataBounds(PropertiesData):
 
         self._set_component('part_node_count', part_node_count, copy=False)
 
-    @_inplace_enabled
+    @_inplace_enabled(default=False)
     def squeeze(self, axes=None, inplace=False):
         '''Remove size one axes from the data array.
 
@@ -1413,7 +1413,7 @@ class PropertiesDataBounds(PropertiesData):
 
         return c
 
-    @_inplace_enabled
+    @_inplace_enabled(default=False)
     def transpose(self, axes=None, inplace=False):
         '''Permute the axes of the data array.
 
@@ -1509,7 +1509,7 @@ class PropertiesDataBounds(PropertiesData):
 
         return c
 
-    @_inplace_enabled
+    @_inplace_enabled(default=False)
     def uncompress(self, inplace=False):
         '''Uncompress the construct.
 
