+++ cfdm/field.py
@@ -501,7 +501,7 @@ class Field(mixin.NetCDFVariable,
 
     @_test_decorator_args('i')
     @_manage_log_level_via_verbosity
-    @_inplace_enabled
+    @_inplace_enabled(default=False)
     def _test_docstring_substitution_Field(self, inplace=False, verbose=None):
         '''Test docstring substitution on {{class}} with two decorators.
 
@@ -721,7 +721,7 @@ class Field(mixin.NetCDFVariable,
 
         return out
 
-    @_inplace_enabled
+    @_inplace_enabled(default=False)
     def compress(self, method, axes=None, count_properties=None,
                  index_properties=None, list_properties=None,
                  inplace=False):
@@ -1471,7 +1471,7 @@ class Field(mixin.NetCDFVariable,
 
         return False
 
-    @_inplace_enabled
+    @_inplace_enabled(default=False)
     def insert_dimension(self, axis, position=0, inplace=False):
         '''Expand the shape of the data array.
 
@@ -2485,7 +2485,7 @@ class Field(mixin.NetCDFVariable,
         for v in variables:
             v.nc_clear_sample_dimension_groups()
 
-    @_inplace_enabled
+    @_inplace_enabled(default=False)
     def squeeze(self, axes=None, inplace=False):
         '''Remove size one axes from the data array.
 
@@ -2557,7 +2557,7 @@ class Field(mixin.NetCDFVariable,
 
         return f
 
-    @_inplace_enabled
+    @_inplace_enabled(default=False)
     def transpose(self, axes=None, constructs=False, inplace=False):
         '''Permute the axes of the data array.
 
@@ -2657,7 +2657,7 @@ class Field(mixin.NetCDFVariable,
 
         return f
 
-    @_inplace_enabled
+    @_inplace_enabled(default=False)
     def uncompress(self, inplace=False):
         '''Uncompress the field construct.
 
