### 2020-09-07 4394999
updated *[get|set|has|del]_data docs
