### 2020-09-07 c4aeb0f
updated *[get|set|has|del]_data docs
