### 2020-09-07 cc16963
updated *[get|set|has|del]_data docs
