+++ cfdm/core/abstract/propertiesdata.py
@@ -211,7 +211,7 @@ class PropertiesData(Properties):
     >>> f.data
     <{{repr}}Data(3): [1, 2, 3]>
     >>> f.del_data()
-    <{{repr}} Data(3): [1, 2, 3]>
+    <{{repr}}Data(3): [1, 2, 3]>
     >>> g = f.set_data([4, 5, 6], inplace=False)
     >>> g.data
     <{{repr}}Data(3): [4, 5, 6]>
@@ -262,7 +262,7 @@ class PropertiesData(Properties):
     >>> f.data
     <{{repr}}Data(3): [1, 2, 3]>
     >>> f.del_data()
-    <{{repr}} Data(3): [1, 2, 3]>
+    <{{repr}}Data(3): [1, 2, 3]>
     >>> g = f.set_data([4, 5, 6], inplace=False)
     >>> g.data
     <{{repr}}Data(3): [4, 5, 6]>
@@ -355,7 +355,7 @@ class PropertiesData(Properties):
     >>> f.data
     <{{repr}}Data(3): [1, 2, 3]>
     >>> f.del_data()
-    <{{repr}} Data(3): [1, 2, 3]>
+    <{{repr}}Data(3): [1, 2, 3]>
     >>> g = f.set_data([4, 5, 6], inplace=False)
     >>> g.data
     <{{repr}}Data(3): [4, 5, 6]>
@@ -412,7 +412,7 @@ class PropertiesData(Properties):
     >>> f.data
     <{{repr}}Data(3): [1, 2, 3]>
     >>> f.del_data()
-    <{{repr}} Data(3): [1, 2, 3]>
+    <{{repr}}Data(3): [1, 2, 3]>
     >>> g = f.set_data([4, 5, 6], inplace=False)
     >>> g.data
     <{{repr}}Data(3): [4, 5, 6]>
