+++ CODE_OF_CONDUCT.md
@@ -50,24 +50,10 @@ they deem inappropriate, threatening, offensive, or harmful.
 This Code of Conduct applies within all project spaces, and it also
 applies when an individual is representing the project or its
 community in public spaces. Examples of representing a project or
-community include using an official project e-mail address, posting
-via an official social media account, or acting as an appointed
-representative at an online or offline event. Representation of a
-project may be further defined and clarified by project maintainers.
-
-## Enforcement
-
-Instances of abusive, harassing, or otherwise unacceptable behavior
-may be reported by contacting the project team at [INSERT EMAIL
-ADDRESS]. All complaints will be reviewed and investigated and will
-result in a response that is deemed necessary and appropriate to the
-circumstances. The project team is obligated to maintain
-confidentiality with regard to the reporter of an incident.  Further
-details of specific enforcement policies may be posted separately.
-
-Project maintainers who do not follow or enforce the Code of Conduct
-in good faith may face temporary or permanent repercussions as
-determined by other members of the project's leadership.
+community include posting via an official social media account, or
+acting as an appointed representative at an online or offline
+event. Representation of a project may be further defined and
+clarified by project maintainers.
 
 ## Attribution
 
