+++ .github/ISSUE_TEMPLATE/question.md
@@ -1,0 +1,11 @@
+---
+name: Question
+about: 'Ask a question, such as "how can I do this?", "why does it behave like that?",
+  "how can I make it faster?", etc. '
+title: 'Question: '
+labels: question
+assignees: ''
+
+---
+
+
