+++ docs/source/contributing.rst
@@ -1,10 +1,14 @@
 .. currentmodule:: cfdm
 .. default-role:: obj
 
-.. _Contribution-guide:
+.. _Contributing:
 
-**Contribution guide**
-======================
+**Contributing**
+================
+
+----
+
+Version |release| for version |version| of the CF conventions.
 
 .. contents::
    :local:
@@ -17,6 +21,8 @@ Please report bugs via a new issue in issue tracker
 (https://github.com/NCAS-CMS/cfdm/issues), using the **Bug report**
 issue template.
 
+----
+
 **Feature requests and suggested improvements**
 -----------------------------------------------
 
@@ -26,6 +32,8 @@ via a new issue in issue tracker
 (https://github.com/NCAS-CMS/cfdm/issues), using the **Feature
 request** issue template.
 
+----
+
 **Questions**
 -------------
 
@@ -33,6 +41,8 @@ Question, such as "how can I do this?", "why does it behave like
 that?", "how can I make it faster?", etc., can be raised via a new
 issue in issue tracker (https://github.com/NCAS-CMS/cfdm/issues),
 using the **Question** issue template.
+
+----
 
 **Preparing pull requests**
 ---------------------------
@@ -58,27 +68,25 @@ Break your edits up into reasonably sized commits.
 	  
     $ git commit -a -m "<COMMIT MESSAGE>"
 
-Run all the tests
+Create a new changelog entry in ``Changelog.rst``. The entry should be
+entered (where ``<description>`` is a brief description of the change)
+as:
+
+.. code-block:: rst
+
+   * <description> (https://github.com/NCAS-CMS/cfdm/issues/<issue number>)
+
+Run the test suite:
 	
 .. code-block:: console
 
    $ cd cfdm/test
    $ python run_tests.py
 
-Create a new changelog entry in Changelog.rst. The entry should be
-entered as:
+Add your name to the list of contributors list at
+``docs/source/contributing.rst``
 
-.. code-block:: rst
-
-   * <description> (https://github.com/NCAS-CMS/cfdm/issues/<issue
-     number>)
-
-where <description> is a brief description of the change.
-
-Add yourself to the list of contributors list at
-``docs/source/contributoring.rst``
-
-Finally, make sure all commits have been puched to the remote copy of
+Finally, make sure all commits have been pushed to the remote copy of
 your fork and submit the pull request via the GitHub website, to the
 ``master`` branch of the ``NCAS-CMS/cfdm`` repository. Make sure to
 reference the original issue in the pull request's description.
@@ -86,16 +94,19 @@ reference the original issue in the pull request's description.
 Note that you can create the pull request while you're working on
 this, as it will automatically update as you add more commits.
 
+----
+
 **Contributors**
 ----------------
 
 We would like to acknowledge and thank all those who have contributed
-ideas and code to the cfdm library:
+ideas, code and documentation to the cfdm library:
 
 * Allyn Treshansky
 * Bryan Lawrence
 * David Hassell
 * Jonathan Gregory
 * Martin Juckes
+* Riley Brady  
 * Sadie Bartholomew  
 
