### 2020-09-08 46fdb79
PEP8 compliance: test Python files outside of 'cfdm/' dir
