### 2020-09-08 23ad09b
remove bare 'bare except' cases
