+++ cfdm/test/test_style.py
@@ -27,7 +27,6 @@ class styleTest(unittest.TestCase):
             'E272',  # ...>1 spaces to align keywords in long import listings
             'E402',  # ...justified lower module imports in {.., core}/__init__
             'E501',  # ...docstring examples include output lines >79 chars
-            'E722',  # ...lots of "bare except" cases need to be addressed
         )
 
         # Find all Python source code ('.py') files in the 'cfdm' directory,
