### 2020-09-08 47c5de9
PEP8 (files outside of cfdm/):: Sphinx docs conf.py now passes
