+++ docs/source/conf.py
@@ -16,12 +16,17 @@ import datetime
 import sys
 import os
 import re
+import inspect
+
+from os.path import relpath, dirname
+
 import cfdm
 
 print('\ncfdm environment:')
 print('-----------------')
 cfdm.environment()
 print()
+
 
 def _read(fname):
     '''Returns content of a file.
@@ -32,6 +37,7 @@ def _read(fname):
     with open(fpath, 'r') as _file:
         return _file.read()
 
+
 def _get_version():
     '''Returns library version by inspecting core/__init__.py file.
 
@@ -40,11 +46,13 @@ def _get_version():
                      _read("../../cfdm/core/__init__.py"),
                      re.MULTILINE).group(1)
 
+
 def _get_cf_version():
     '''Returns CF version by inspecting core/__init__.py file.
 
     '''
     return cfdm.__cf_version__
+
 
 def _get_year():
     '''Get the current calendar year.
@@ -52,11 +60,13 @@ def _get_year():
     '''
     return str(datetime.datetime.now().year)
 
+
 def _get_date():
     '''Get the current calendar year.
 
     '''
     return str(datetime.date.today())
+
 
 # If extensions (or modules to document with autodoc) are in another
 # directory, add these directories to sys.path here. If the directory
@@ -70,26 +80,27 @@ sys.path.insert(0, os.path.abspath('../..'))
 # If your documentation needs a minimal Sphinx version, state it here.
 needs_sphinx = '2.3.1'
 
-#rst_prolog = """
-#.. |CF| replace:: """+_get_cf_version()+"""
-#"""
+# rst_prolog = """
+# .. |CF| replace:: """+_get_cf_version()+"""
+# """
 
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
-extensions = ['sphinx.ext.autodoc',
-              'sphinx.ext.autosummary',
-#              'sphinx.ext.viewcode',
-              'sphinx.ext.linkcode',
-              'sphinx.ext.mathjax',
-              'sphinx.ext.graphviz',
-#              'sphinx.ext.inheritance_diagram',
-              'sphinx.ext.intersphinx',
-              'sphinx.ext.doctest',
-              'sphinx.ext.githubpages',
-#              'sphinxcontrib.programoutput',  # pip install sphinxcontrib-programoutput
-              'sphinx_copybutton',
-              'sphinx_toggleprompt',
+extensions = [
+    'sphinx.ext.autodoc',
+    'sphinx.ext.autosummary',
+    # 'sphinx.ext.viewcode',
+    'sphinx.ext.linkcode',
+    'sphinx.ext.mathjax',
+    'sphinx.ext.graphviz',
+    # 'sphinx.ext.inheritance_diagram',
+    'sphinx.ext.intersphinx',
+    'sphinx.ext.doctest',
+    'sphinx.ext.githubpages',
+    # 'sphinxcontrib.programoutput',  # pip install sphinxcontrib-programoutput
+    'sphinx_copybutton',
+    'sphinx_toggleprompt',
 ]
 
 
@@ -102,9 +113,9 @@ autosummary_generate = True
 # and inserted.
 autoclass_content = 'both'
 
-#inheritance_graph_attrs = {'rankdir': "TB",
-#                           'clusterrank': 'local'}
-#inheritance_node_attrs  = {'style': 'filled'}
+# inheritance_graph_attrs = {'rankdir': "TB",
+#                            'clusterrank': 'local'}
+# inheritance_node_attrs  = {'style': 'filled'}
 
 # This value selects how automatically documented members are sorted
[…]
