### 2020-09-08 99a7575
docstring rewriter docs
