+++ docs/source/class/cfdm.Index.rst
@@ -152,4 +152,17 @@ Special
    ~cfdm.Index.__repr__
    ~cfdm.Index.__str__
 
-   
+Docstring substitutions                      
+-----------------------                   
+                                          
+.. rubric:: Methods                       
+                                          
+.. autosummary::                          
+   :nosignatures:                         
+   :toctree: ../method/                   
+   :template: method.rst                  
+                                          
+   ~cfdm.Index._docstring_special_substitutions
+   ~cfdm.Index._docstring_substitutions        
+   ~cfdm.Index._docstring_package_depth        
+   ~cfdm.Index._docstring_method_exclusions    
