+++ docs/source/class/cfdm.core.DomainAncillary.rst
@@ -112,4 +112,17 @@ Special
 
    ~cfdm.core.DomainAncillary.__deepcopy__
 
-   
+Docstring substitutions                           
+-----------------------                        
+                                               
+.. rubric:: Methods                            
+                                               
+.. autosummary::                               
+   :nosignatures:                              
+   :toctree: ../method/                        
+   :template: method.rst                       
+                                               
+   ~cfdm.core.DomainAncillary._docstring_special_substitutions
+   ~cfdm.core.DomainAncillary._docstring_substitutions        
+   ~cfdm.core.DomainAncillary._docstring_package_depth        
+   ~cfdm.core.DomainAncillary._docstring_method_exclusions    
