+++ docs/source/class/cfdm.core.meta.DocstringRewriteMeta.rst
@@ -13,6 +13,8 @@ cfdm.core.meta.DocstringRewriteMeta
 Inspection
 ----------
 
+.. rubric:: Methods
+
 .. autosummary::
    :nosignatures:
    :toctree: ../method/
