+++ docs/source/class/cfdm.core.InteriorRing.rst
@@ -76,3 +76,17 @@ Special
 
    ~cfdm.core.InteriorRing.__deepcopy__
 
+Docstring substitutions                   
+-----------------------                   
+                                          
+.. rubric:: Methods                       
+                                          
+.. autosummary::                          
+   :nosignatures:                         
+   :toctree: ../method/                   
+   :template: method.rst                  
+                                          
+   ~cfdm.core.InteriorRing._docstring_special_substitutions
+   ~cfdm.core.InteriorRing._docstring_substitutions        
+   ~cfdm.core.InteriorRing._docstring_package_depth        
+   ~cfdm.core.InteriorRing._docstring_method_exclusions    
