+++ docs/source/class/cfdm.core.Field.rst
@@ -155,3 +155,18 @@ Special
    :template: method.rst
 
    ~cfdm.core.Field.__deepcopy__
+
+Docstring substitutions                           
+-----------------------                        
+                                               
+.. rubric:: Methods                            
+                                               
+.. autosummary::                               
+   :nosignatures:                              
+   :toctree: ../method/                        
+   :template: method.rst                       
+                                               
+   ~cfdm.core.Field._docstring_special_substitutions
+   ~cfdm.core.Field._docstring_substitutions        
+   ~cfdm.core.Field._docstring_package_depth        
+   ~cfdm.core.Field._docstring_method_exclusions    
