+++ docs/source/class/cfdm.Array.rst
@@ -65,3 +65,18 @@ Special
    ~cfdm.Array.__getitem__
    ~cfdm.Array.__repr__
    ~cfdm.Array.__str__
+
+Docstring substitutions                      
+-----------------------                   
+                                          
+.. rubric:: Methods                       
+                                          
+.. autosummary::                          
+   :nosignatures:                         
+   :toctree: ../method/                   
+   :template: method.rst                  
+                                          
+   ~cfdm.Array._docstring_special_substitutions
+   ~cfdm.Array._docstring_substitutions        
+   ~cfdm.Array._docstring_package_depth        
+   ~cfdm.Array._docstring_method_exclusions    
