+++ docs/source/class/cfdm.core.abstract.PropertiesData.rst
@@ -76,3 +76,18 @@ Special
    :template: method.rst
 
    ~cfdm.core.abstract.PropertiesData.__deepcopy__
+
+Docstring substitutions                                    
+-----------------------                                 
+                                                        
+.. rubric:: Methods                                     
+                                                        
+.. autosummary::                                        
+   :nosignatures:                                       
+   :toctree: ../method/                                 
+   :template: method.rst                                
+                                                        
+   ~cfdm.core.abstract.PropertiesData._docstring_special_substitutions
+   ~cfdm.core.abstract.PropertiesData._docstring_substitutions        
+   ~cfdm.core.abstract.PropertiesData._docstring_package_depth        
+   ~cfdm.core.abstract.PropertiesData._docstring_method_exclusions    
