### 2020-09-08 8d79f75
Changelog: note discovered bug fix (we might forget at release-time)
