+++ Changelog.rst
@@ -1,3 +1,12 @@
+version 1.8.7.0
+---------------
+----
+
+**2020-??-??**
+
+* Fixed bug that prevented the writing of multiple fields to netCDF when
+  at least one dimension was shared between some of the fields.
+
 version 1.8.6.0
 ---------------
 ----
