### 2020-09-08 59b6116
Catch only specific Exception cases in conf.py
