+++ docs/source/conf.py
@@ -477,7 +477,7 @@ def linkcode_resolve(domain, info):
     for part in fullname.split('.'):
         try:
             obj = getattr(obj, part)
-        except:
+        except AttributeError:
             return None
 
     # Get the object wrapped by obj. This makes sure that the actual
@@ -486,19 +486,19 @@ def linkcode_resolve(domain, info):
 
     try:
         fn = inspect.getsourcefile(obj)
-    except:
+    except TypeError:
         fn = None
     if not fn:
         return None
 
     try:
         source, lineno = inspect.findsource(obj)
-    except:
+    except OSError:
         lineno = None
 
     try:
         nlines = len(inspect.getsourcelines(obj)[0])
-    except:
+    except (OSError, IndexError):
         nlines = None
 
     fn = relpath(fn, start=dirname(cfdm.__file__))
