### 2020-09-09 30011b7
Docstring rewriter test: hard-coding -> generation of subclasses
