### 2020-09-09 e3d6ab1
Docstring rewriter: fix two issues picked up by revamped test
