### 2020-09-10 5d979c0
JOSS paper text: remove spurious semicolons and commas
