### 2020-09-10 129f297
JOSS paper text: consistent use of serial comma (always use)
