### 2020-09-10 a409840
JOSS paper text: improvements to some wording
