### 2020-09-10 aa4e285
JOSS paper text: clarify sentence in statement of need
