+++ cfdm/core/meta/docstringrewrite.py
@@ -522,6 +522,27 @@ class DocstringRewriteMeta(type):
         # Do general substitutions first
         # ------------------------------------------------------------
         for key, value in config.items():
+            
+            for k, v in config.items():
+                if k in value:
+                    try:
+                        # Compiled regular expression substitution
+                        value = key.sub(v, value)
+                    except AttributeError:
+                        # String substitution
+                        doc = value.replace(k, v)
+            # --- End: for
+                
+            try:
+                # Compiled regular expression substitution
+                doc = key.sub(value, doc)
+            except AttributeError:
+                # String substitution
+                doc = doc.replace(key, value)
+                
+        # --- End: for
+
+        for key, value in config.items():
             try:
                 # Compiled regular expression substitution
                 doc = key.sub(value, doc)
