### 2020-09-11 f298ca5
allow docstring substitutions to contain other non-special docstring substitutions
