+++ cfdm/fieldancillary.py
@@ -53,8 +53,6 @@ class FieldAncillary(mixin.NetCDFVariable,
 
         {{init data: data_like, optional}}
 
-            {{data_like}}
-
         source: optional
             Initialize the properties and data from those of *source*.
 
