+++ cfdm/dimensioncoordinate.py
@@ -59,8 +59,6 @@ class DimensionCoordinate(mixin.NetCDFVariable,
 
         {{init data: data_like, optional}}
 
-            {{data_like}}
-
         {{init bounds: `Bounds`, optional}}
 
         {{init geometry: `str`, optional}}
