+++ cfdm/auxiliarycoordinate.py
@@ -58,8 +58,6 @@ class AuxiliaryCoordinate(mixin.NetCDFVariable,
 
         {{init data: data_like, optional}}
 
-            {{data_like}}
-
         {{init bounds: `Bounds`, optional}}
 
         {{init geometry: `str`, optional}}
