+++ cfdm/list.py
@@ -40,8 +40,6 @@ class List(mixin.NetCDFVariable,
 
         {{init data: data_like, optional}}
 
-            {{data_like}}
-
         source: optional
             Initialize the properties and data from those of *source*.
 
