### 2020-09-14 #2531 @davidhassell
@bradyrx, @usethedata

Apologies, I inserted an incorrect version of the "New text" at the end of the previous comments. it should read:

**(New Text)** "To the authors’ knowledge, cfdm and software built on it are currently the only libraries that can understand all CF-netCDF datasets, made possible by the complete implementation of the CF data model. All others omit facets that are not currently of interest to their particular user communities."

This is what is in the latest `paper.md` and pdf.

Thanks
