### 2020-09-18 5eb8b89
GH Actions: upload coverage reports to Codecov for two repr. jobs
