### 2020-09-18 726b765
Close #78: update docstrings of methods to get constructs &/or keys
