### 2020-09-18 9b0b9f2
GH Actions: remove shell key which breaks 'upload coverage...' step
