+++ .github/workflows/run-test-suite.yml
@@ -102,7 +102,6 @@ jobs:
       if: |
         (matrix.os == 'ubuntu-latest' || matrix.os == 'macos-latest') &&
         matrix.python-version == 3.7
-      shell: bash -l {0}
       with:
         file: ./coverage.xml
         directory: ./cfdm_coverage_reports
