### 2020-09-18 #2531 @bradyrx
Wonderful @davidhassell. Thanks so much for the prompt response and incorporating my suggestions. Looks good to me. I am good for this to be accepted. I know we're still on a `PRE REVIEW` thread, but hopefully this does it. Congratulations, and nice work on the package. That's quite the headache to handle all aspects of CF but very important!
