### 2020-09-21 6bba2e0
GH Actions: fix step to upload coverage reports
* path to coverage XML was incorrect
* latest @1.X version must be specified (is not taken by default),
  see: https://github.com/codecov/codecov-action/issues/43
