+++ cfdm/constructs.py
@@ -14,14 +14,33 @@ class Constructs(mixin.Container,
                  core.Constructs):
     '''A container for metadata constructs.
 
+    The following metadata constructs can be included:
+
+    * auxiliary coordinate constructs
+    * coordinate reference constructs
+    * cell measure constructs
+    * dimension coordinate constructs
+    * domain ancillary constructs
+    * domain axis constructs
+    * cell method constructs
+    * field ancillary constructs
+
+    The container may be used by `Field` and `Domain` instances. In
+    the latter case cell method and field ancillary constructs must be
+    flagged as "ignored" (see the *_ignore* parameter).
+
+    The container is like a dictionary in many ways, in that it stores
+    key/value pairs where the key is the unique construct key with
+    correspondaing metadata construct value, and provides some of the
+    usual dictionary methods.
+
+
+    **Calling**
+
     Calling a `Constructs` instance selects metadata constructs by
-    identity and is an alias for the `filter_by_identity` method.
-
-    **Examples:**
-
-    Select constructs that have a ``standard_name`` property of 'foo':
-
-    >>> d = c('foo')
+    identity and is an alias for the `filter_by_identity` method. For
+    example, to select constructs that have an identity of
+    'air_temperature': ``d = c('air_temperature')``.
 
     .. versionadded:: (cfdm) 1.7.0
 
@@ -30,7 +49,9 @@ class Constructs(mixin.Container,
         '''Select metadata constructs by identity.
 
     Calling a `Constructs` instance selects metadata constructs by
-    identity and is an alias for the `filter_by_identity` method.
+    identity and is an alias for the `filter_by_identity` method. For
+    example, to select constructs that have an identity of
+    'air_temperature': ``d = c('air_temperature')``.
 
     .. versionadded:: (cfdm) 1.7.0
 
