+++ docs/source/class/cfdm.DomainAncillary.rst
@@ -1,6 +1,8 @@
 .. currentmodule:: cfdm
 .. default-role:: obj
 
+.. _cfdm-DomainAncillary:
+
 cfdm.DomainAncillary
 ====================
 
