### 2020-09-22 6ac7f1b
Domain inherits from Properties
