+++ docs/source/class/cfdm.core.Domain.rst
@@ -38,6 +38,24 @@ Metadata constructs
 
    ~cfdm.core.Domain.constructs
    
+Properties
+----------
+
+.. rubric:: Methods
+	    
+.. autosummary::
+   :nosignatures:
+   :toctree: ../method/
+   :template: method.rst
+
+   ~cfdm.core.Domain.del_property
+   ~cfdm.core.Domain.get_property
+   ~cfdm.core.Domain.has_property
+   ~cfdm.core.Domain.set_property
+   ~cfdm.core.Domain.properties
+   ~cfdm.core.Domain.clear_properties
+   ~cfdm.core.Domain.set_properties
+
 Miscellaneous
 -------------
 
