+++ docs/source/installation.rst
@@ -70,7 +70,7 @@ To install from source:
 1. Download the cfdm package from https://pypi.org/project/cfdm
 
 2. Unpack the library (replacing ``<version>`` with the version that
-   you want to install, e.g. ``1.8.6.0``):
+   you want to install, e.g. ``1.9.0.0``):
 
    .. code:: console
 	 
