### 2020-09-22 8cbd97e
Domain properties unit test
