+++ cfdm/read_write/netcdf/netcdfread.py
@@ -710,6 +710,12 @@ class NetCDFRead(IORead):
             g['version'][version] = LooseVersion(version)
 
         # ------------------------------------------------------------
+        # Modify read vars if we are only reading domain variables
+        # ------------------------------------------------------------
+        if g['domain']:
+            g['mask'] = False
+
+        # ------------------------------------------------------------
         # Add custom read vars
         # ------------------------------------------------------------
         if extra_read_vars:
@@ -717,9 +723,9 @@ class NetCDFRead(IORead):
 
         compression = {}
 
-        # ----------------------------------------------------------------
+        # ------------------------------------------------------------
         # Parse field parameter
-        # ----------------------------------------------------------------
+        # ------------------------------------------------------------
         g['get_constructs'] = {
             'auxiliary_coordinate': self.implementation.get_auxiliary_coordinates,
             'cell_measure': self.implementation.get_cell_measures,
@@ -825,9 +831,11 @@ class NetCDFRead(IORead):
 
         g['file_version'] = LooseVersion(file_version)
 
-        # Set minimum versions
+        # Set minimum/maximum versions
         for vn in ('1.6', '1.7', '1.8', '1.9'):
             g['CF>='+vn] = (g['file_version'] >= g['version'][vn])
+
+        g['CF<=1.8'] = not g['CF>=1.9']
 
         # ------------------------------------------------------------
         # Create a dictionary keyed by netCDF variable names where
@@ -1357,13 +1365,15 @@ class NetCDFRead(IORead):
         # --- End: if
 
         # ------------------------------------------------------------
-        # Create a field from every netCDF variable (apart from
+        # Create a field/domain from every netCDF variable (apart from
         # special variables that have already been identified as such)
         # ------------------------------------------------------------
         all_fields = OrderedDict()
         for ncvar in g['variables']:
             if ncvar not in g['do_not_create_field']:
-                all_fields[ncvar] = self._create_field(ncvar)
+                field_or_domain = self._create_field(ncvar)
+                if field_or_domain is not None:
+                    all_fields[ncvar] = field_or_domain
         # --- End: for
 
         # ------------------------------------------------------------
@@ -1394,8 +1404,8 @@ class NetCDFRead(IORead):
         )  # pragma: no cover
 
         # ------------------------------------------------------------
-        # Discard fields created from netCDF variables that are
-        # referenced by other netCDF variables
+        # Discard fields/domains created from netCDF variables that
+        # are referenced by other netCDF variables
         # ------------------------------------------------------------
         fields = OrderedDict()
         for ncvar, f in all_fields.items():
@@ -1438,7 +1448,7 @@ class NetCDFRead(IORead):
         # that are referenced by other netCDF variables.
         # ------------------------------------------------------------
         self_referenced = {}
-        if g['extra']:
+        if g['extra'] and not g['domain']:
             fields0 = list(fields.values())
             for construct_type in g['extra']:
                 for f in fields0:
@@ -1473,7 +1483,7 @@ class NetCDFRead(IORead):
                     x.dataset_compliance(display=True)
         # --- End: if
 
-        if warn_valid:
+        if warn_valid and not g['domain']:
             # --------------------------------------------------------
             # Warn for the presence of 'valid_min', 'valid_max'or
             # 'valid_range' properties. (Introduced at v1.8.3)
@@ -1494,7 +1504,7 @@ class NetCDFRead(IORead):
         self.file_close()
 
         # ------------------------------------------------------------
-        # Return the fields
+        # Return the fields/domains
         # ------------------------------------------------------------
         return out
 
@@ -2733,6 +2743,8 @@ class NetCDFRead(IORead):
         '''
         g = self.read_vars
 
+        create_field = g['CF<=1.8'] or not g['domain']
+
         # Reset 'domain_ancillary_key'
[…]
