+++ cfdm/abstract/implementation.py
@@ -41,6 +41,7 @@ class Implementation(metaclass=abc.ABCMeta):
      'Data',
      'Datum',
      'DimensionCoordinate',
+     'Domain',
      'DomainAncillary',
      'DomainAxis',
      'Field',
