### 2020-09-23 faecc90
moved methods to constructaccess.py, fixed some domain read issues
