+++ cfdm/read_write/netcdf/netcdfread.py
@@ -710,12 +710,6 @@ class NetCDFRead(IORead):
             g['version'][version] = LooseVersion(version)
 
         # ------------------------------------------------------------
-        # Modify read vars if we are only reading domain variables
-        # ------------------------------------------------------------
-        if g['domain']:
-            g['mask'] = False
-
-        # ------------------------------------------------------------
         # Add custom read vars
         # ------------------------------------------------------------
         if extra_read_vars:
@@ -834,8 +828,6 @@ class NetCDFRead(IORead):
         # Set minimum/maximum versions
         for vn in ('1.6', '1.7', '1.8', '1.9'):
             g['CF>='+vn] = (g['file_version'] >= g['version'][vn])
-
-        g['CF<=1.8'] = not g['CF>=1.9']
 
         # ------------------------------------------------------------
         # Create a dictionary keyed by netCDF variable names where
@@ -1371,7 +1363,8 @@ class NetCDFRead(IORead):
         all_fields = OrderedDict()
         for ncvar in g['variables']:
             if ncvar not in g['do_not_create_field']:
-                field_or_domain = self._create_field(ncvar)
+                field_or_domain = self._create_field_or_domain(
+                    ncvar, domain=g['domain'])
                 if field_or_domain is not None:
                     all_fields[ncvar] = field_or_domain
         # --- End: for
@@ -1418,7 +1411,7 @@ class NetCDFRead(IORead):
         unreferenced_variables = [ncvar for ncvar in sorted(all_fields)
                                   if self._is_unreferenced(ncvar)]
 
-        xx = []
+#        xx = []
         for ncvar in referenced_variables[:]:
             if all(referencer in referenced_variables
                    for referencer in g['referencers'][ncvar]):
@@ -1485,8 +1478,9 @@ class NetCDFRead(IORead):
 
         if warn_valid and not g['domain']:
             # --------------------------------------------------------
-            # Warn for the presence of 'valid_min', 'valid_max'or
-            # 'valid_range' properties. (Introduced at v1.8.3)
+            # Warn for the presence of field 'valid_min',
+            # 'valid_max'or 'valid_range' properties. (Introduced at
+            # v1.8.3)
             # --------------------------------------------------------
             for f in out:
                 # Check field constructs
@@ -2726,8 +2720,8 @@ class NetCDFRead(IORead):
 
         return ncvar, message
 
-    def _create_field(self, field_ncvar):
-        '''Create a field for a given netCDF variable.
+    def _create_field_or_domain(self, field_ncvar, domain=False):
+        '''Create a field or domain for a given netCDF variable.
 
     .. versionadded:: (cfdm) 1.7.0
 
@@ -2736,6 +2730,12 @@ class NetCDFRead(IORead):
         field_ncvar: `str`
             The name of the netCDF variable to be turned into a field.
 
+        domain: `bool`, otpional
+            If True then only read and parse domain variables. By
+            default only data variables are read and parsed.
+
+            .. versionadded:: (cfdm) 1.9.0.0
+
     :Returns:
 
         Field construct
@@ -2743,10 +2743,11 @@ class NetCDFRead(IORead):
         '''
         g = self.read_vars
 
-        create_field = g['CF<=1.8'] or not g['domain']
-
-# TODO NEED TO ONLY MAKE DOMAIN VARIABELS FROM UNREFERENCED VARIBLES ???
+        field = not domain
+               
         
+# TODO NEED TO ONLY MAKE DOMAIN VARIABELS FROM UNREFERENCED VARIBLES ???
+
         # Reset 'domain_ancillary_key'
         g['domain_ancillary_key'] = {}
 
@@ -2759,9 +2760,14 @@ class NetCDFRead(IORead):
[…]
