+++ cfdm/field.py
@@ -348,74 +348,74 @@ class Field(mixin.NetCDFVariable,
     # ----------------------------------------------------------------
     # Private methods
     # ----------------------------------------------------------------
-    def _get_data_compression_variables(self, component):
-        '''
-
-        '''
-        out = []
-        for construct in self.constructs.filter_by_data().values():
-            data = construct.get_data(None)
-            if data is None:
-                continue
-
-            x = getattr(data, 'get_' + component)(None)
-            if x is None:
-                continue
-
-            out.append(x)
-
-        for construct in self.constructs.filter_by_data().values():
-            if not construct.has_bounds():
-                continue
-
-            data = construct.get_bounds_data(None)
-            if data is None:
-                continue
-
-            x = getattr(data, 'get_' + component)(None)
-            if x is None:
-                continue
-
-            out.append(x)
-
-        for construct in self.coordinates.values():
-            interior_ring = construct.get_interior_ring(None)
-            if interior_ring is None:
-                continue
-
-            data = interior_ring.get_data(None)
-            if data is None:
-                continue
-
-            x = getattr(data, 'get_' + component)(None)
-            if x is None:
-                continue
-
-            out.append(x)
-
-        return out
-
-    def _get_coordinate_geometry_variables(self, component):
-        '''Return the list of variables for the geometry coordinates.
-
-    :Parameters:
-
-        component: `str`
-
-    :Returns:
-
-        `list'
-
-        '''
-        out = []
-        for construct in self.coordinates.values():
-            x = getattr(construct, 'get_' + component)(None)
-            if x is None:
-                continue
-
-            out.append(x)
-
-        return out
+#    def _get_data_compression_variables(self, component):
+#        '''
+#        
+#        '''
+#        out = []
+#        for construct in self.constructs.filter_by_data().values():
+#            data = construct.get_data(None)
+#            if data is None:
+#                continue
+#
+#            x = getattr(data, 'get_' + component)(None)
+#            if x is None:
+#                continue
+#
+#            out.append(x)
+#
+#        for construct in self.constructs.filter_by_data().values():
+#            if not construct.has_bounds():
+#                continue
+#
+#            data = construct.get_bounds_data(None)
+#            if data is None:
+#                continue
+#
+#            x = getattr(data, 'get_' + component)(None)
+#            if x is None:
+#                continue
+#
+#            out.append(x)
+#
+#        for construct in self.coordinates.values():
+#            interior_ring = construct.get_interior_ring(None)
+#            if interior_ring is None:
+#                continue
+#
+#            data = interior_ring.get_data(None)
+#            if data is None:
+#                continue
+#
+#            x = getattr(data, 'get_' + component)(None)
+#            if x is None:
+#                continue
+#
+#            out.append(x)
+#
+#        return out
+#
+#    def _get_coordinate_geometry_variables(self, component):
+#        '''Return the list of variables for the geometry coordinates.
+#
+#    :Parameters:
+#
+#        component: `str`
+#
+#    :Returns:
+#
+#        `list'
+#
+#        '''
+#        out = []
+#        for construct in self.coordinates.values():
+#            x = getattr(construct, 'get_' + component)(None)
+#            if x is None:
+#                continue
+#
+#            out.append(x)
+#
+#        return out
 
     def _one_line_description(self, axis_names_sizes=None):
         '''
@@ -1791,700 +1791,700 @@ class Field(mixin.NetCDFVariable,
             print('    },')
             print('}\n')
 
-    def nc_set_component_variable(self, component, value):
-        '''Set the netCDF variable name for all components of the given type.
-
-    Some components exist within multiple constructs, but when written
[…]
