### 2020-09-23 1543fb1
Domain: properties, cfdm.write
