+++ cfdm/mixin/constructaccess.py
@@ -88,7 +88,7 @@ class ConstructAccess():
 
     def _get_data_compression_variables(self, component):
         '''
-        
+
         '''
         out = []
         for construct in self.constructs.filter_by_data().values():
