+++ cfdm/read_write/netcdf/netcdfwrite.py
@@ -2440,7 +2440,8 @@ class NetCDFWrite(IOWrite):
 
     def _write_netcdf_variable(self, ncvar, ncdimensions, cfvar,
                                omit=(), extra=None, fill=False,
-                               data_variable=False):
+                               data_variable=False,
+                               domain_variable=False):
         '''Create a netCDF variable from *cfvar* with name *ncvar* and
     dimensions *ncdimensions*. The new netCDF variable's properties
     are given by cfvar.properties(), less any given by the *omit*
@@ -2463,6 +2464,14 @@ class NetCDFWrite(IOWrite):
 
         extra: `dict`, optional
 
+        data_variable: `bool`, optional
+            True if the a data variable is being written.
+
+        domain_variable: `bool`, optional
+            True if the a domain variable is being written.
+
+            .. versionadded:: (cfdm) 1.9.0.0
+
     :Returns:
 
         `None`
@@ -2479,13 +2488,19 @@ class NetCDFWrite(IOWrite):
         # ------------------------------------------------------------
         # Set the netCDF4.createVariable datatype
         # ------------------------------------------------------------
-        datatype = self._datatype(cfvar)
-        data = self.implementation.get_data(cfvar, None)
+        if domain_variable:
+            data = None
+            datatype = 'S1'
+            original_ncdimensions = ()
+            ncdimensions = ()
+        else:
+            datatype = self._datatype(cfvar)
+            data = self.implementation.get_data(cfvar, None)
 
-        original_ncdimensions = ncdimensions
+            original_ncdimensions = ncdimensions
 
-        data, ncdimensions = self._transform_strings(cfvar, data,
-                                                     ncdimensions)
+            data, ncdimensions = self._transform_strings(cfvar, data,
+                                                         ncdimensions)
 
         # ------------------------------------------------------------
         # Find the fill value - the value that the variable's data get
@@ -2835,8 +2850,8 @@ class NetCDFWrite(IOWrite):
     def _convert_to_char(self, data):
         '''Convert string data into character data
 
-    The return Data instance object will have data type 'S1' and will
-    have an extra trailing dimension.
+    The returned Data instance object will have data type 'S1' and
+    will have an extra trailing dimension.
 
     .. versionadded:: (cfdm) 1.7.0
 
@@ -2850,8 +2865,7 @@ class NetCDFWrite(IOWrite):
 
         '''
         strlen = data.dtype.itemsize
-#        print ('1 strlen =', strlen)
-#        if strlen > 1:
+
         data = self.implementation.initialise_Data(
                 array=self._character_array(
                     self.implementation.get_array(data)),
@@ -2861,15 +2875,17 @@ class NetCDFWrite(IOWrite):
 
         return data
 
-    def _write_field(self, f, add_to_seen=False,
-                     allow_data_insert_dimension=True):
-        '''TODO
+    def _write_field_or_domain(self, f, add_to_seen=False,
+                               allow_data_insert_dimension=True):
+        '''Write a field or domain construct to the file.
+
+    All of the metadata constructs are also written.
 
     .. versionadded:: (cfdm) 1.7.0
 
     :Parameters:
 
-        f: Field construct
+        f: `Field` or `Domain`
 
         add_to_seen: bool, optional
 
@@ -2884,6 +2900,16 @@ class NetCDFWrite(IOWrite):
 
         logger.info('  Writing {!r}:'.format(f))  # pragma: no cover
 
+        # Work out if we have a field or domain instance. CF-1.9
+        domain = isinstance(f, self.implementation.get_class('Domain'))
+        field = not domain
+
+        if domain and g['output_version'] < g['CF-1.9']:
+            raise ValueError(
[…]
