+++ cfdm/mixin/properties.py
@@ -318,8 +318,8 @@ class Properties(Container):
     The identities comprise:
 
     * The ``standard_name`` property.
-    * All properties, preceeded by the property name and a colon,
-      e.g. ``'long_name:Air temperature'``.
+    * All properties, preceeded by the property name and an equals
+      e.g. ``'long_name=Air temperature'``.
     * The netCDF variable name, preceeded by ``'ncvar%'``.
 
     .. versionadded:: (cfdm) 1.7.0
