### 2020-09-23 a405d9f
transfer of creation_commands methods from cf-python
