+++ cfdm/__init__.py
@@ -109,7 +109,6 @@ from .decorators import (
     _manage_log_level_via_verbosity,
 )
 
-
 from .constructs import Constructs
 
 from .data import (Data,
