+++ cfdm/cellmethod.py
@@ -122,15 +122,9 @@ class CellMethod(mixin.Container,
         '''
         namespace0 = namespace
         if namespace is None:
-            namespace = self._namespace() + '.'
+            namespace = self._package() + '.'
         elif namespace and not namespace.endswith('.'):
             namespace += '.'
-
-#        namespace0 = namespace
-#        if namespace0:
-#            namespace = namespace+"."
-#       else:
-#             namespace = ""
 
         indent = ' ' * indent
 
