+++ cfdm/coordinatereference.py
@@ -191,15 +191,9 @@ class CoordinateReference(mixin.NetCDFVariable,
         '''
         namespace0 = namespace
         if namespace is None:
-            namespace = self._namespace() + '.'
+            namespace = self._package() + '.'
         elif namespace and not namespace.endswith('.'):
             namespace += '.'
-
-#        namespace0 = namespace
-#        if namespace0:
-#            namespace = namespace+"."
-#        else:
-#            namespace = ""
 
         indent = ' ' * indent
 
