+++ cfdm/data/data.py
@@ -1185,7 +1185,7 @@ class Data(Container,
         '''
         namespace0 = namespace
         if namespace is None:
-            namespace = self._namespace() + '.'
+            namespace = self._package() + '.'
         elif namespace and not namespace.endswith('.'):
             namespace += '.'
 
