+++ cfdm/mixin/propertiesdatabounds.py
@@ -574,7 +574,7 @@ class PropertiesDataBounds(PropertiesData):
 
         namespace0 = namespace
         if namespace is None:
-            namespace = self._namespace() + '.'
+            namespace = self._package() + '.'
         elif namespace and not namespace.endswith('.'):
             namespace += '.'
 
