+++ cfdm/mixin/propertiesdata.py
@@ -479,7 +479,7 @@ class PropertiesData(Properties):
 
         namespace0 = namespace
         if namespace is None:
-            namespace = self._namespace() + '.'
+            namespace = self._package() + '.'
         elif namespace and not namespace.endswith('.'):
             namespace += '.'
 
