+++ docs/source/class/cfdm.Field.rst
@@ -175,6 +175,7 @@ Miscellaneous
 
    ~cfdm.Field.compress
    ~cfdm.Field.copy
+   ~cfdm.Field.creation_commands
    ~cfdm.Field.equals
    ~cfdm.Field.convert
    ~cfdm.Field.uncompress
