+++ docs/source/tutorial.rst
@@ -2174,6 +2174,12 @@ There are three methods for creating a field construct in memory:
   provide the connections between them.
 ..
 
+* :ref:`Command modification <Command-modification>`: Produce the
+  commands that would create an already existing field construct, and
+  then modify them.
+
+..
+
 * :ref:`Creation by conversion <Creation-by-conversion>`: Convert a
   single metadata construct already in memory to an independent field
   construct
@@ -2626,7 +2632,96 @@ The new field construct may now be inspected:
    Domain ancils   : domainancillary0(atmosphere_hybrid_height_coordinate(1)) = [10.0] m
                    : domainancillary1(atmosphere_hybrid_height_coordinate(1)) = [20.0] 1
                    : surface_altitude(grid_latitude(10), grid_longitude(9)) = [[0.0, ..., 89.0]] m
-		  
+	
+.. _Command-modification:
+
+Command modification
+^^^^^^^^^^^^^^^^^^^^
+
+It is sometimes convenient to produce the commands that would create
+an already existing field construct, and then modify them to create
+the desired field construct. The commands are produced by the
+`~Field.creation_commands` method of the existing field construct.
+
+.. code-block:: python
+   :caption: *Produce the commands that would create an existing field
+             construct.*
+	
+   >>> q, t = cfdm.read('file.nc')
+   >>> print(q.creation_commands())
+   # field: specific_humidity
+   f = cfdm.Field()
+   f.set_properties({'Conventions': 'CF-1.7', 'project': 'research', 'standard_name': 'specific_humidity', 'units': '1'})
+   d = cfdm.Data([[0.007, 0.034, 0.003, 0.014, 0.018, 0.037, 0.024, 0.029], [0.023, 0.036, 0.045, 0.062, 0.046, 0.073, 0.006, 0.066], [0.11, 0.131, 0.124, 0.146, 0.087, 0.103, 0.057, 0.011], [0.029, 0.059, 0.039, 0.07, 0.058, 0.072, 0.009, 0.017], [0.006, 0.036, 0.019, 0.035, 0.018, 0.037, 0.034, 0.013]], units='1', dtype='f8')
+   f.set_data(d)
+   f.nc_set_variable('q')
+   
+   # netCDF global attributes
+   f.nc_set_global_attributes({'Conventions': None, 'project': None})
+   
+   # domain_axis: ncdim%lat
+   c = cfdm.DomainAxis()
+   c.set_size(5)
+   c.nc_set_dimension('lat')
+   f.set_construct(c, key='domainaxis0', copy=False)
+   
+   # domain_axis: ncdim%lon
+   c = cfdm.DomainAxis()
+   c.set_size(8)
+   c.nc_set_dimension('lon')
+   f.set_construct(c, key='domainaxis1', copy=False)
+   
+   # domain_axis: 
+   c = cfdm.DomainAxis()
+   c.set_size(1)
+   f.set_construct(c, key='domainaxis2', copy=False)
+   
+   # field data axes
+   f.set_data_axes(('domainaxis0', 'domainaxis1'))
+   
+   # dimension_coordinate: latitude
+   c = cfdm.DimensionCoordinate()
+   c.set_properties({'units': 'degrees_north', 'standard_name': 'latitude'})
+   d = cfdm.Data([-75.0, -45.0, 0.0, 45.0, 75.0], units='degrees_north', dtype='f8')
+   c.set_data(d)
+   c.nc_set_variable('lat')
+   b = cfdm.Bounds()
[…]
