+++ Changelog.rst
@@ -13,6 +13,8 @@ version 1.8.7.0
 * New method: `cfdm.Field._docstring_package_depth`
 * New method: `cfdm.Field._docstring_method_exclusions`
 * New function: `cfdm.unique_domains`
+* Enabled "creation commands" methods
+  (https://github.com/NCAS-CMS/cfdm/issues/53)
 * Fixed bug that caused failures when reading or writing a dataset
   that contains multiple geometry containers
   (https://github.com/NCAS-CMS/cfdm/issues/65)
