+++ cfdm/domain.py
@@ -389,7 +389,8 @@ class Domain(mixin.NetCDFVariable,
 
     def creation_commands(self, representative_data=False,
                           namespace=None, indent=0, string=True,
-                          name='f'):
+                          name='domain', data_name='data',
+                          header=True):
         '''Return the commands that would create the domain construct.
 
     **Construct keys**
@@ -418,156 +419,29 @@ class Domain(mixin.NetCDFVariable,
 
         {{string: `bool`, optional}}
 
+        {{header: `bool`, optional}}
+
     :Returns:
 
         {{returns creation_commands}}
 
     **Examples:**
 
-    >>> q = {{package}}.example_field(0)
-    >>> print(q)
-    Field: specific_humidity (ncvar%q)
-    ----------------------------------
-    Data            : specific_humidity(latitude(5), longitude(8)) 1
-    Cell methods    : area: mean
-    Dimension coords: latitude(5) = [-75.0, ..., 75.0] degrees_north
-                    : longitude(8) = [22.5, ..., 337.5] degrees_east
-                    : time(1) = [2019-01-01 00:00:00]
-    >>> print(q.creation_commands())
-    # field: specific_humidity
-    f = {{package}}.Field()
-    #
-    f.set_properties({'Conventions': 'CF-1.7', 'project': 'research', 'standard_name': 'specific_humidity', 'units': '1'})
-    f.nc_set_variable('q')
-    #
-    # domain_axis
-    c = {{package}}.DomainAxis(size=5)
-    c.nc_set_dimension('lat')
-    f.set_construct(c, key='domainaxis0')
-    #
-    # domain_axis
-    c = {{package}}.DomainAxis(size=8)
-    c.nc_set_dimension('lon')
-    f.set_construct(c, key='domainaxis1')
-    #
-    # domain_axis
-    c = {{package}}.DomainAxis(size=1)
-    f.set_construct(c, key='domainaxis2')
-    #
-    # field data
-    data = {{package}}.Data([[0.007, 0.034, 0.003, 0.014, 0.018, 0.037, 0.024, 0.029], [0.023, 0.036, 0.045, 0.062, 0.046, 0.073, 0.006, 0.066], [0.11, 0.131, 0.124, 0.146, 0.087, 0.103, 0.057, 0.011], [0.029, 0.059, 0.039, 0.07, 0.058, 0.072, 0.009, 0.017], [0.006, 0.036, 0.019, 0.035, 0.018, 0.037, 0.034, 0.013]], units='1', dtype='f8')
-    f.set_data(data, axes=('domainaxis0', 'domainaxis1'))
-    #
-    # dimension_coordinate
-    c = {{package}}.DimensionCoordinate()
-    c.set_properties({'units': 'degrees_north', 'standard_name': 'latitude'})
-    c.nc_set_variable('lat')
-    data = {{package}}.Data([-75.0, -45.0, 0.0, 45.0, 75.0], units='degrees_north', dtype='f8')
-    c.set_data(data)
-    b = {{package}}.Bounds()
-    b.nc_set_variable('lat_bnds')
-    data = {{package}}.Data([[-90.0, -60.0], [-60.0, -30.0], [-30.0, 30.0], [30.0, 60.0], [60.0, 90.0]], units='degrees_north', dtype='f8')
-    b.set_data(data)
-    c.set_bounds(b)
-    f.set_construct(c, axes=('domainaxis0',), key='dimensioncoordinate0', copy=False)
-    #
-    # dimension_coordinate
-    c = {{package}}.DimensionCoordinate()
[…]
