+++ cfdm/field.py
@@ -1720,6 +1720,257 @@ class Field(mixin.NetCDFVariable,
 
         return f
 
+    def creation_commands(self, representative_data=False,
+                          namespace=None, indent=0, string=True,
+                          name='field', data_name='data', header=True):
+        '''Return the commands that would create the field construct.
+
+    **Construct keys**
+
+    The *key* parameter of the output `set_construct` commands is
+    utilised in order minimise the number of commands needed to
+    implement cross-referencing between constructs (e.g. between a
+    coordinate reference construct and coordinate constructs). This is
+    usually not necessary when building field constructs, as by
+    default the `set_construct` method returns a unique construct key
+    for the construct being set.
+
+    .. versionadded:: (cfdm) 1.8.7.0
+
+    .. seealso:: `set_construct`,
+                 `{{package}}.Data.creation_commands`,
+                 `{{package}}.example_field`
+
+    :Parameters:
+
+        {{representative_data: `bool`, optional}}
+
+        {{namespace: `str`, optional}}
+
+        {{indent: `int`, optional}}
+
+        {{string: `bool`, optional}}
+
+        {{header: `bool`, optional}}
+
+    :Returns:
+
+        {{returns creation_commands}}
+
+    **Examples:**
+
+    >>> q = {{package}}.example_field(0)
+    >>> print(q)
+    Field: specific_humidity (ncvar%q)
+    ----------------------------------
+    Data            : specific_humidity(latitude(5), longitude(8)) 1
+    Cell methods    : area: mean
+    Dimension coords: latitude(5) = [-75.0, ..., 75.0] degrees_north
+                    : longitude(8) = [22.5, ..., 337.5] degrees_east
+                    : time(1) = [2019-01-01 00:00:00]
+    >>> print(q.creation_commands())
+    #
+    # field: specific_humidity
+    field = cfdm.Field()
+    field.set_properties({'Conventions': 'CF-1.8', 'project': 'research', 'standard_name': 'specific_humidity', 'units': '1'})
+    field.nc_set_variable('q')
+    data = cfdm.Data([[0.007, 0.034, 0.003, 0.014, 0.018, 0.037, 0.024, 0.029], [0.023, 0.036, 0.045, 0.062, 0.046, 0.073, 0.006, 0.066], [0.11, 0.131, 0.124, 0.146, 0.087, 0.103, 0.057, 0.011], [0.029, 0.059, 0.039, 0.07, 0.058, 0.072, 0.009, 0.017], [0.006, 0.036, 0.019, 0.035, 0.018, 0.037, 0.034, 0.013]], units='1', dtype='f8')
+    field.set_data(data)
+    #
+    # domain_axis: ncdim%lat
+    c = cfdm.DomainAxis()
+    c.set_size(5)
+    c.nc_set_dimension('lat')
+    field.set_construct(c, key='domainaxis0', copy=False)
+    #
+    # domain_axis: ncdim%lon
+    c = cfdm.DomainAxis()
+    c.set_size(8)
+    c.nc_set_dimension('lon')
+    field.set_construct(c, key='domainaxis1', copy=False)
+    #
+    # domain_axis:
+    c = cfdm.DomainAxis()
+    c.set_size(1)
+    field.set_construct(c, key='domainaxis2', copy=False)
+    #
+    # dimension_coordinate: latitude
+    c = cfdm.DimensionCoordinate()
+    c.set_properties({'units': 'degrees_north', 'standard_name': 'latitude'})
+    c.nc_set_variable('lat')
[…]
