+++ docs/source/class/cfdm.DimensionCoordinate.rst
@@ -109,6 +109,21 @@ Bounds
 
    ~cfdm.DimensionCoordinate.bounds
 
+Climatological time
+^^^^^^^^^^^^^^^^^^^
+
+.. rubric:: Methods
+	    
+.. autosummary::
+   :nosignatures:
+   :toctree: ../method/
+   :template: method.rst
+
+   ~cfdm.DimensionCoordinate.del_climatology
+   ~cfdm.DimensionCoordinate.get_climatology
+   ~cfdm.DimensionCoordinate.is_climatology
+   ~cfdm.DimensionCoordinate.set_climatology
+
 Miscellaneous
 -------------
 
