+++ docs/source/class/cfdm.core.DimensionCoordinate.rst
@@ -88,6 +88,21 @@ Bounds
 
    ~cfdm.core.DimensionCoordinate.bounds
 
+Climatological time
+^^^^^^^^^^^^^^^^^^^
+
+.. rubric:: Methods
+	    
+.. autosummary::
+   :nosignatures:
+   :toctree: ../method/
+   :template: method.rst
+
+   ~cfdm.core.DimensionCoordinate.del_climatology
+   ~cfdm.core.DimensionCoordinate.get_climatology
+   ~cfdm.core.DimensionCoordinate.is_climatology
+   ~cfdm.core.DimensionCoordinate.set_climatology
+
 Miscellaneous
 -------------
 
