+++ .github/workflows/run-test-suite.yml
@@ -98,7 +98,7 @@ jobs:
     - name: Upload coverage report to Codecov
       # Get coverage for the two latest distros but only one Python version to
       # treat as representative (probably overkill to get on any more jobs).
-      uses: codecov/codecov-action@v1.0.7
+      uses: codecov/codecov-action@v1.0.13
       if: |
         (matrix.os == 'ubuntu-latest' || matrix.os == 'macos-latest') &&
         matrix.python-version == 3.7
