+++ cfdm/data/gatheredarray.py
@@ -131,7 +131,11 @@ class GatheredArray(abstract.CompressedArray):
 
     :Parameters:
 
-        {{default: optional}}
+        default: optional
+            Return the value of the *default* parameter if the list
+            variable has not been set.
+
+            {{default Exception}}
 
     :Returns:
 
