+++ cfdm/test/test_Field.py
@@ -617,24 +617,23 @@ class FieldTest(unittest.TestCase):
 
         for rd in (False, True):
             for indent in (0, 4):
-                for s in (False, True):
-                    for ns in ('cfdm', ''):
-                        _ = f.creation_commands(representative_data=rd,
-                                                indent=indent,
-                                                namespace=ns,
-                                                string=s)
-        # --- End: for
-
-        for i in range(7):
-            f = cfdm.example_field(i)
-            for rd in (False, True):
-                for indent in (0, 4):
+                for h in (False, True):
                     for s in (False, True):
-                        for ns in ('cfdm', ''):
-                            _ = f.creation_commands(representative_data=rd,
-                                                    indent=indent,
-                                                    namespace=ns,
-                                                    string=s)
+                        for ns in (None, ''):
+                            _ = f.creation_commands(
+                                representative_data=rd,
+                                indent=indent,
+                                namespace=ns,
+                                string=s,
+                                header=h)
+                            for i in range(7):
+                                f = cfdm.example_field(i)
+                                _ = f.creation_commands(
+                                    representative_data=rd,
+                                    indent=indent,
+                                    namespace=ns,
+                                    string=s,
+                                    header=h)
         # --- End: for
 
 # --- End: class
