### 2020-09-25 55ff722
Docs: add 'key%' prefixes in tutorial example for exact output match
