### 2020-09-25 3068b48
Docs via docstrings: apply one spelling fix throughout codebase
