### 2020-09-26 ebc63ad
ordered constructs corner case
