+++ cfdm/test/test_read_write.py
@@ -61,7 +61,7 @@ class read_writeTest(unittest.TestCase):
         # self.test_only = ['NOTHING!!!!!']
         # self.test_only = ['test_write_filename']
         # self.test_only = ['test_read_write_unlimited']
-        # self.test_only = ['test_read_field']
+        self.test_only = ['test_read_write_domain']
         # self.test_only = ['test_read_mask']
         # self.test_only = ['test_read_write_format']
         # self.test_only = ['test_read_write_Conventions']
@@ -444,13 +444,13 @@ class read_writeTest(unittest.TestCase):
 
         tmpfile = 'delme.nc'
         cfdm.write(d, tmpfile)
-        e = cfdm.read(tmpfile, domain=False)
-        self.assertFalse(e)
+        e = cfdm.read(tmpfile)
+        self.assertTrue(len(e), 10)
         
-        e = cfdm.read(tmpfile, domain=True)
+        e = cfdm.read(tmpfile, domain=True, verbose=-1)
         self.assertEqual(len(e), 1)
         e = e[0]
-        self.assertIsInsstance(e, cfdm.Domain)
+        self.assertIsInstance(e, cfdm.Domain)
                 
         raise ValueError('Write some more tests!')
 # --- End: class
