+++ cfdm/field.py
@@ -442,27 +442,27 @@ class Field(mixin.NetCDFVariable,
 
         return "{0}{1}{2}".format(self.identity(''), axis_names, units)
 
-    def _set_dataset_compliance(self, value):
-        '''Set the report of problems encountered whilst reading the field
-    construct from a dataset.
-
-    .. versionadded:: (cfdm) 1.7.0
-
-    .. seealso:: `dataset_compliance`
-
-    :Parameters:
-
-        value:
-           The value of the ``dataset_compliance`` component.
-
-    :Returns:
-
-        `None`
-
-    **Examples:**
-
-        '''
-        self._set_component('dataset_compliance', value, copy=True)
+#    def _set_dataset_compliance(self, value):
+#        '''Set the report of problems encountered whilst reading the field
+#    construct from a dataset.
+#
+#    .. versionadded:: (cfdm) 1.7.0
+#
+#    .. seealso:: `dataset_compliance`
+#
+#    :Parameters:
+#
+#        value:
+#           The value of the ``dataset_compliance`` component.
+#
+#    :Returns:
+#
+#        `None`
+#
+#    **Examples:**
+#
+#        '''
+#        self._set_component('dataset_compliance', value, copy=True)
 
     @property
     def _test_docstring_substitution_property_Field(self):
