+++ cfdm/field.py
@@ -1220,6 +1220,7 @@ class Field(mixin.NetCDFVariable,
 
     .. seealso:: `set_construct`,
                  `{{package}}.Data.creation_commands`,
+                 `{{package}}.Domain.creation_commands`,
                  `{{package}}.example_field`
 
     :Parameters:
@@ -1392,39 +1393,36 @@ class Field(mixin.NetCDFVariable,
         out = super().creation_commands(
             representative_data=representative_data, indent=0,
             namespace=namespace, string=False, name=name,
-            data_name=data_name, header=header)
+            data_name=data_name, header=header
+        )
 
         nc_global_attributes = self.nc_global_attributes()
         if nc_global_attributes:
             out.append('#')
             out.append('# netCDF global attributes')
             out.append("{}.nc_set_global_attributes({!r})".format(
-                name, nc_global_attributes))
-
-        # Domain axes
-        for key, c in self.domain_axes.items():
-            out.extend(
-                c.creation_commands(
-                    indent=0, string=False,
-                    namespace=namespace0, name='c',
-                    header=header)
+                name, nc_global_attributes)
             )
-            out.append("{}.set_construct(c, key={!r}, copy=False)".format(
-                name, key))
+
+        # Domain
+        out.extend(
+            self.domain.creation_commands(
+                representative_data=representative_data,
+                string=False, indent=0, namespace=namespace0,
+                name=name, data_name=data_name, header=header,
+                _domain=False
+            )
+        )
 
         # Metadata constructs with data
-        for key, c in self.constructs.filter_by_type(
-                'dimension_coordinate',
-                'auxiliary_coordinate',
-                'cell_measure',
-                'domain_ancillary',
-                'field_ancillary').items():
+        for key, c in self.field_ancillaries().items():
             out.extend(
                 c.creation_commands(
                     representative_data=representative_data, string=False,
                     indent=0, namespace=namespace0, name='c',
                     data_name=data_name,
-                    header=header)
+                    header=header
+                )
             )
             out.append(
                 "{}.set_construct(c, axes={}, key={!r}, copy=False)".format(
@@ -1436,17 +1434,8 @@ class Field(mixin.NetCDFVariable,
                 c.creation_commands(namespace=namespace0,
                                     indent=0, string=False,
                                     name='c',
-                                    header=header)
-            )
-            out.append("{}.set_construct(c)".format(name))
-
-        # Coordinate reference constructs
-        for key, c in self.coordinate_references.items():
-            out.extend(
-                c.creation_commands(namespace=namespace0,
-                                    indent=0, string=False,
-                                    name='c',
-                                    header=header)
+                                    header=header
+                )
             )
             out.append("{}.set_construct(c)".format(name))
 
@@ -1966,327 +1955,327 @@ class Field(mixin.NetCDFVariable,
 
         return f
 
[…]
