+++ cfdm/field.py
@@ -1700,13 +1700,13 @@ class Field(mixin.NetCDFVariable,
 
         return out
 
-    def has_geometry(self):
-        '''TODO'''
-        for c in self.coordinates.values():
-            if c.has_geometry():
-                return True
-
-        return False
+#    def has_geometry(self):
+#        '''TODO'''
+#        for c in self.coordinates.values():
+#            if c.has_geometry():
+#                return True
+#
+#        return False
 
     @_inplace_enabled(default=False)
     def insert_dimension(self, axis, position=0, inplace=False):
