+++ cfdm/core/cellmeasure.py
@@ -49,13 +49,10 @@ class CellMeasure(abstract.PropertiesData):
         {{init data: data_like, optional}}
 
         source: optional
-            Initialise the *measure*, *properties* and *data*
-            parameters (if present) from *source*, which will be a
-            `CellMeasure` object, or a subclass of one of its parent
-            classes.
+            Initialise the measure, properties and data from those of
+            source.
 
-            *Parameter example:*
-              >>> d = CellMeasure(source=c)
+            {{init source}}
 
         {{init copy: `bool`, optional}}
 
@@ -86,7 +83,8 @@ class CellMeasure(abstract.PropertiesData):
 
     **Examples:**
 
-    >>> f.construct_type
+    >>> c = {{package}}.{{class}}()
+    >>> c.construct_type
     'cell_measure'
 
         '''
