+++ cfdm/constructs.py
@@ -436,7 +436,7 @@ class Constructs(mixin.Container,
 
     :Returns:
 
-        `{{class}}`
+        `Constructs`
             The deep copy.
 
     **Examples:**
@@ -847,7 +847,7 @@ class Constructs(mixin.Container,
 
     :Returns:
 
-        `{{class}}`
+        `Constructs`
             The selected constructs and their construct keys.
 
     **Examples:**
@@ -971,7 +971,7 @@ class Constructs(mixin.Container,
 
     :Returns:
 
-        `{{class}}`
+        `Constructs`
             The selected constructs and their construct keys.
 
     **Examples:**
@@ -1042,7 +1042,7 @@ class Constructs(mixin.Container,
 
     :Returns:
 
-        `{{class}}`
+        `Constructs`
             The selected constructs and their construct keys.
 
     **Examples:**
@@ -1121,7 +1121,7 @@ class Constructs(mixin.Container,
 
     :Returns:
 
-        `{{class}}`
+        `Constructs`
             The selected constructs and their construct keys.
 
     **Examples:**
@@ -1179,7 +1179,7 @@ class Constructs(mixin.Container,
 
     :Returns:
 
-        `{{class}}`
+        `Constructs`
             The selected cell measure constructs and their construct
             keys.
 
@@ -1279,7 +1279,7 @@ class Constructs(mixin.Container,
 
     :Returns:
 
-        `{{class}}`
+        `Constructs`
             The selected cell method constructs and their construct
             keys.
 
@@ -1379,7 +1379,7 @@ class Constructs(mixin.Container,
 
     :Returns:
 
-        `{{class}}`
+        `Constructs`
             The selected domain axis constructs and their construct
             keys.
 
@@ -1458,7 +1458,7 @@ class Constructs(mixin.Container,
 
     :Returns:
 
-        `{{class}}`
+        `Constructs`
             The selected domain axis constructs and their construct
             keys.
 
@@ -1539,7 +1539,7 @@ class Constructs(mixin.Container,
 
     :Returns:
 
-        `{{class}}`
+        `Constructs`
             The selected constructs and their construct keys.
 
     **Examples:**
@@ -1646,7 +1646,7 @@ class Constructs(mixin.Container,
 
     :Returns:
 
-        `{{class}}`
+        `Constructs`
             The selected constructs and their construct keys.
 
     **Examples:**
@@ -1748,7 +1748,7 @@ class Constructs(mixin.Container,
 
     :Returns:
 
-        `{{class}}`
+        `Constructs`
             The selected domain axis constructs and their construct
             keys.
 
@@ -1830,7 +1830,7 @@ class Constructs(mixin.Container,
 
     :Returns:
 
-        `{{class}}`
+        `Constructs`
             The selected constructs and their construct keys.
 
     **Examples:**
@@ -1996,7 +1996,7 @@ class Constructs(mixin.Container,
 
     :Returns:
 
-        `{{class}}`
+        `Constructs`
             The constructs, and their construct keys, that were not
             selected by the last filter applied. If no filtering has
             been applied, or the last filter was an inverse filter,
@@ -2091,7 +2091,7 @@ class Constructs(mixin.Container,
 
     :Returns:
 
-        `{{class}}`
+        `Constructs`
             The shallow copy.
 
     **Examples:**
@@ -2143,7 +2143,7 @@ class Constructs(mixin.Container,
 
     :Returns:
 
-        `{{class}}`
+        `Constructs`
             The constructs, and their construct keys, that existed
             before the last filter was applied. If no filters have
             been applied then all of the constructs are returned.
