+++ cfdm/test/test_Data.py
@@ -55,7 +55,6 @@ class DataTest(unittest.TestCase):
         self.assertFalse(d.any())
 
     def test_Data__repr__str(self):
-        '''Check cfdm.Data.__repr__, cfdm.Data.__str__'''
         if self.test_only and inspect.stack()[0][3] not in self.test_only:
             return
 
@@ -80,7 +79,6 @@ class DataTest(unittest.TestCase):
 #            return
 
     def test_Data__setitem__(self):
-        '''Check cfdm.Data.__setitem__'''
         if self.test_only and inspect.stack()[0][3] not in self.test_only:
             return
 
@@ -140,7 +138,6 @@ class DataTest(unittest.TestCase):
         self.assertIs(a[()], numpy.ma.masked)
 
     def test_Data_apply_masking(self):
-        '''Check cfdm.Data.apply_masking'''
         if self.test_only and inspect.stack()[0][3] not in self.test_only:
             return
 
@@ -218,7 +215,6 @@ class DataTest(unittest.TestCase):
 #            pass
 
     def test_Data_array(self):
-        '''Check cfdm.Data.array'''
         if self.test_only and inspect.stack()[0][3] not in self.test_only:
             return
 
@@ -263,7 +259,6 @@ class DataTest(unittest.TestCase):
         self.assertFalse((a2 == a).all())
 
     def test_Data_datetime_array(self):
-        '''Check cfdm.Data.datetime_array'''
         if self.test_only and inspect.stack()[0][3] not in self.test_only:
             return
 
@@ -286,7 +281,6 @@ class DataTest(unittest.TestCase):
         self.assertIs(dt[()], numpy.ma.masked)
 
     def test_Data_flatten(self):
-        '''Check cfdm.Data.flatten'''
         if self.test_only and inspect.stack()[0][3] not in self.test_only:
             return
 
@@ -324,7 +318,6 @@ class DataTest(unittest.TestCase):
             self.assertEqual(e.size, d.size)
 
     def test_Data_transpose(self):
-        '''Check cfdm.Data.transpose'''
         if self.test_only and inspect.stack()[0][3] not in self.test_only:
             return
 
@@ -342,7 +335,6 @@ class DataTest(unittest.TestCase):
         # --- End: for
 
     def test_Data_unique(self):
-        '''Check cfdm.Data.unique'''
         if self.test_only and inspect.stack()[0][3] not in self.test_only:
             return
 
@@ -359,7 +351,6 @@ class DataTest(unittest.TestCase):
             (u.array == cfdm.Data([1, 2, 4], 'metre').array).all())
 
     def test_Data_equals(self):
-        '''Check cfdm.Data.equals'''
         if self.test_only and inspect.stack()[0][3] not in self.test_only:
             return
 
@@ -374,8 +365,6 @@ class DataTest(unittest.TestCase):
         self.assertTrue(e.equals(d, verbose=3))
 
     def test_Data_max_min_sum_squeeze(self):
-        '''Check cfdm.Data.maximum, cfdm.Data.minumum, cfdm.Data.sum,
-    cfdm.Data.squeeze'''
         if self.test_only and inspect.stack()[0][3] not in self.test_only:
             return
 
@@ -421,7 +410,6 @@ class DataTest(unittest.TestCase):
         self.assertTrue((x.array == b).all(), (x.shape, b.shape))
 
     def test_Data_dtype_mask(self):
-        '''Check cfdm.Data.dtype, cfdm.Data.mask'''
         if self.test_only and inspect.stack()[0][3] not in self.test_only:
             return
 
@@ -480,6 +468,22 @@ class DataTest(unittest.TestCase):
[…]
