+++ cfdm/mixin/propertiesdatabounds.py
@@ -551,6 +551,24 @@ class PropertiesDataBounds(PropertiesData):
 
         {{string: `bool`, optional}}
 
+        {{name: `str`, optional}}
+
+        {{data_name: `str`, optional}}
+
+        bounds_name: `str`, optional
+            The name of the construct's `Bounds` instance created by
+            the returned commands.
+
+            *Parameter example:*
+              ``name='bounds1'``
+
+        interior_ring_name: `str`, optional
+            The name of the construct's `InteriorRing` instance
+            created by the returned commands.
+
+            *Parameter example:*
+              ``name='ir1'``
+
         {{header: `bool`, optional}}
 
     :Returns:
@@ -559,7 +577,23 @@ class PropertiesDataBounds(PropertiesData):
 
     **Examples:**
 
-        TODO
+    >>> x = {{package}}.{{class}}(
+    ...     properties={'units': 'degrees_east',
+    ...                 'standard_name': 'longitude'}
+    ... )
+    >>> x.set_data([22.5, 67.5, 112.5])
+    >>> b = {{package}}.Bounds()
+    >>> b.set_data([[0.0, 45.0], [45.0, 90.0], [90.0, 135.0]])
+    >>> x.set_bounds(b)
+    >>> print(x.creation_commands(header=False))
+    c = {{package}}.{{class}}()
+    c.set_properties({'units': 'degrees_east', 'standard_name': 'longitude'})
+    data = {{package}}.Data([22.5, 67.5, 112.5], units='degrees_east', dtype='f8')
+    c.set_data(data)
+    b = {{package}}.Bounds()
+    data = {{package}}.Data([[0.0, 45.0], [45.0, 90.0], [90.0, 135.0]], units='degrees_east', dtype='f8')
+    b.set_data(data)
+    c.set_bounds(b)
 
         '''
         if name in (data_name, bounds_name, interior_ring_name):

[…]
