### 2020-10-01 dfe48a9
squeeze scalar data
