+++ cfdm/test/test_Data.py
@@ -313,15 +313,30 @@ class DataTest(unittest.TestCase):
                     numpy.prod([n
                                 for i, n in enumerate(d.shape) if i in axes]))
 
-            self.assertEqual(e.shape, tuple(shape))
+            self.assertEqual(e.shape, tuple(shape), axes)
             self.assertEqual(e.ndim, d.ndim-len(axes)+1)
             self.assertEqual(e.size, d.size)
+
+        for n in range(4):
+            e = d.flatten(n)
+            f = d.flatten([n])
+            self.assertTrue(e.equals(f))
+
+        with self.assertRaises(ValueError):
+            d.flatten(99)
+
+        d = cfdm.Data(9)
+        self.assertTrue(d.equals(d.flatten()))
+        self.assertTrue(d.equals(d.flatten([])))
+
+        with self.assertRaises(ValueError):
+            d.flatten(0)
 
     def test_Data_transpose(self):
         if self.test_only and inspect.stack()[0][3] not in self.test_only:
             return
 
-        a = numpy.arange(10*15*19).reshape(10, 1, 15, 19)
+        a = numpy.arange(2 * 3 * 5).reshape(2, 1, 3, 5)
         d = cfdm.Data(a.copy())
 
         for indices in (list(range(a.ndim)), list(range(-a.ndim, 0))):
@@ -333,6 +348,15 @@ class DataTest(unittest.TestCase):
                 self.assertEqual(d.shape, a.shape, message)
                 self.assertTrue((d.array == a).all(), message)
         # --- End: for
+
+        with self.assertRaises(ValueError):
+            d.transpose(axes=99)
+
+        with self.assertRaises(ValueError):
+            d.transpose(axes=[1, 2, 3, 4, 5])
+
+        d = cfdm.Data(9)
+        self.assertTrue(d.equals(d.transpose()))
 
     def test_Data_unique(self):
         if self.test_only and inspect.stack()[0][3] not in self.test_only:
@@ -364,14 +388,14 @@ class DataTest(unittest.TestCase):
         self.assertTrue(d.equals(e, verbose=3))
         self.assertTrue(e.equals(d, verbose=3))
 
-    def test_Data_max_min_sum_squeeze(self):
+    def test_Data_maximum_minimum_sum_squeeze(self):
         if self.test_only and inspect.stack()[0][3] not in self.test_only:
             return
 
-        a = numpy.ma.arange(10*15*19).reshape(10, 1, 15, 19)
+        a = numpy.ma.arange(2 * 3 * 5).reshape(2, 1, 3, 5)
         a[0, 0, 0, 0] = numpy.ma.masked
         a[-1, -1, -1, -1] = numpy.ma.masked
-        a[5, -1, 6, 10] = numpy.ma.masked
+        a[1, -1, 1, 3] = numpy.ma.masked
         d = cfdm.Data(a)
 
         b = a.max()
@@ -388,6 +412,7 @@ class DataTest(unittest.TestCase):
         x = d.maximum(axes=[0, 3]).squeeze([0, 3])
         self.assertEqual(x.shape, b.shape)
         self.assertTrue((x.array == b).all())
+        self.assertTrue(d.maximum([0, 3]).equals(d.max([0, 3])))
 
         b = a.min()
         x = d.minimum().squeeze()
@@ -398,6 +423,7 @@ class DataTest(unittest.TestCase):
         x = d.minimum(axes=[0, 3]).squeeze([0, 3])
         self.assertEqual(x.shape, b.shape)
         self.assertTrue((x.array == b).all(), (x.shape, b.shape))
+        self.assertTrue(d.minimum([0, 3]).equals(d.min([0, 3])))
 
         b = a.sum()
         x = d.sum().squeeze()
[…]
