+++ cfdm/core/constructs.py
@@ -7,6 +7,24 @@ from . import abstract
 class Constructs(abstract.Container):
     '''A container for metadata constucts.
 
+    The following metadata constructs can be included:
+
+    * auxiliary coordinate constructs
+    * coordinate reference constructs
+    * cell measure constructs
+    * dimension coordinate constructs
+    * domain ancillary constructs
+    * domain axis constructs
+    * cell method constructs
+    * field ancillary constructs
+
+    The container is used by used by `Field` and `Domain` instances.
+
+    The container is like a dictionary in many ways, in that it stores
+    key/value pairs where the key is the unique construct key with
+    correspondaing metadata construct value, and provides some of the
+    usual dictionary methods.
+
     .. versionadded:: (cfdm) 1.7.0
 
     '''
@@ -86,6 +104,10 @@ class Constructs(abstract.Container):
             metadata constructs are deep copied.
 
         _ignore: sequence of `str`, optional
+            Ignores the given construct types.
+
+            *Parameter example:*
+              ``_ignore=('cell_method', 'field_ancillary')``
 
         '''
         self._ignore = tuple(set(_ignore))
@@ -297,52 +319,52 @@ class Constructs(abstract.Container):
     # ----------------------------------------------------------------
     # Private methods
     # ----------------------------------------------------------------
-    def _default(self, default, message=None):
-        '''Return a value or raise an Exception for a default case.
-
-    .. versionadded:: (cfdm) 1.7.0
-
-    :Parameters:
-
-        default:
-            The value to return, or to raise if set to an `Exception`
-            instance.
-
-        message: `str`, optional
-            The error message to raise with *default* if it is an
-            `Exception` instance.
-
-    :Returns:
-
-        The value of *default* if it is not an `Exception` instance.
-
-    **Examples:**
-
-    >>> f = cfdm.example_field(0)
-    >>> c = f.get_construct('cellmethod0')
-    >>> c
-    <{{repr}}CellMethod: area: mean>
-
-    >>> c._default(AttributeError())  # Raises Exception
-    AttributeError
-    >>> c._default(ValueError(), message="No data")  # Raises Exception
-    ValueError: No data
-    >>> c._default(False)
-    False
-    >>> c._default('Not set')
-    'Not set'
-    >>> c._default(1)
-    1
-
-        '''
-        if isinstance(default, Exception):
-            if message is not None and not default.args:
-                default = copy(default)
-                default.args = (message,)
-
-            raise default
-
-        return default
+#    def _default(self, default, message=None):
+#        '''Return a value or raise an Exception for a default case.
+#
+#    .. versionadded:: (cfdm) 1.7.0
+#
+#    :Parameters:
+#
+#        default:
+#            The value to return, or to raise if set to an `Exception`
+#            instance.
+#
+#        message: `str`, optional
+#            The error message to raise with *default* if it is an
+#            `Exception` instance.
+#
+#    :Returns:
+#
+#        The value of *default* if it is not an `Exception` instance.
+#
+#    **Examples:**
+#
+#    >>> f = cfdm.example_field(0)
+#    >>> c = f.get_construct('cellmethod0')
+#    >>> c
+#    <{{repr}}CellMethod: area: mean>
+#
+#    >>> c._default(AttributeError())  # Raises Exception
+#    AttributeError
+#    >>> c._default(ValueError(), message="No data")  # Raises Exception
+#    ValueError: No data
+#    >>> c._default(False)
+#    False
+#    >>> c._default('Not set')
[…]
