+++ cfdm/data/data.py
@@ -553,11 +553,9 @@ class Data(Container,
     :Parameters:
 
         axes: (sequence of) `int`
-            The axes of the data. May be one of, or a sequence of any
-            combination of zero or more of:
+            The axes of the data.
 
-            * The integer position of a dimension in the data
-              (negative indices allowed).
+            {{axes int examples}}
 
     :Returns:
 
@@ -1637,6 +1635,10 @@ class Data(Container,
     :Parameters:
 
         axes: (sequence of) `int`, optional
+            The axes over which to take the maximum. By default the
+            maximum over all axes is returned.
+
+            {{axes int examples}}
 
     :Returns:
 
@@ -1644,6 +1646,37 @@ class Data(Container,
             Maximum of the data along the specified axes.
 
     **Examples:**
+
+    >>> import numpy
+    >>> d = {{package}}.Data(numpy.arange(24).reshape(1, 2, 3, 4))
+    >>> d
+    <{{repr}}Data(1, 2, 3, 4): [[[[0, ..., 23]]]]>
+    >>> print(d.array)
+    [[[[ 0  1  2  3]
+       [ 4  5  6  7]
+       [ 8  9 10 11]]
+      [[12 13 14 15]
+       [16 17 18 19]
+       [20 21 22 23]]]]
+    >>> e = d.max()
+    >>> e
+    <{{repr}}Data(1, 1, 1, 1): [[[[23]]]]>
+    >>> print(e.array)
+    [[[[23]]]]
+    >>> e = d.max(2)
+    >>> e
+    <{{repr}}Data(1, 2, 1, 4): [[[[8, ..., 23]]]]>
+    >>> print(e.array)
+    [[[[ 8  9 10 11]]
+
+      [[20 21 22 23]]]]
+    >>> e = d.max([-2, -1])
+    >>> e
+    <{{repr}}Data(1, 2, 1, 1): [[[[11, 23]]]]>
+    >>> print(e.array)
+    [[[[11]]
+
+      [[23]]]]
 
         '''
         # Parse the axes. By default flattened input is used.
@@ -1676,6 +1709,10 @@ class Data(Container,
     :Parameters:
 
         axes: (sequence of) `int`, optional
+            The axes over which to take the minimum. By default the
+            minimum over all axes is returned.
+
+            {{axes int examples}}
 
     :Returns:
 
@@ -1683,6 +1720,37 @@ class Data(Container,
             Minimum of the data along the specified axes.
 
     **Examples:**
+
+    >>> import numpy
+    >>> d = {{package}}.Data(numpy.arange(24).reshape(1, 2, 3, 4))
+    >>> d
+    <{{repr}}Data(1, 2, 3, 4): [[[[0, ..., 23]]]]>
+    >>> print(d.array)
+    [[[[ 0  1  2  3]
+       [ 4  5  6  7]
+       [ 8  9 10 11]]
+      [[12 13 14 15]
+       [16 17 18 19]
+       [20 21 22 23]]]]
+    >>> e = d.min()
+    >>> e
+    <{{repr}}Data(1, 1, 1, 1): [[[[0]]]]>
+    >>> print(e.array)
+    [[[[0]]]]
+    >>> e = d.min(2)
+    >>> e
+    <{{repr}}Data(1, 2, 1, 4): [[[[0, ..., 15]]]]>
+    >>> print(e.array)
+    [[[[ 0  1  2  3]]
+
+      [[12 13 14 15]]]]
+    >>> e = d.min([-2, -1])
+    >>> e
+    <{{repr}}Data(1, 2, 1, 1): [[[[0, 12]]]]>
+    >>> print(e.array)
+    [[[[ 0]]
+
+      [[12]]]]
 
         '''
[…]
