+++ cfdm/data/data.py
@@ -1668,14 +1668,12 @@ class Data(Container,
     <{{repr}}Data(1, 2, 1, 4): [[[[8, ..., 23]]]]>
     >>> print(e.array)
     [[[[ 8  9 10 11]]
-
       [[20 21 22 23]]]]
     >>> e = d.max([-2, -1])
     >>> e
     <{{repr}}Data(1, 2, 1, 1): [[[[11, 23]]]]>
     >>> print(e.array)
     [[[[11]]
-
       [[23]]]]
 
         '''
@@ -1742,14 +1740,12 @@ class Data(Container,
     <{{repr}}Data(1, 2, 1, 4): [[[[0, ..., 15]]]]>
     >>> print(e.array)
     [[[[ 0  1  2  3]]
-
       [[12 13 14 15]]]]
     >>> e = d.min([-2, -1])
     >>> e
     <{{repr}}Data(1, 2, 1, 1): [[[[0, 12]]]]>
     >>> print(e.array)
     [[[[ 0]]
-
       [[12]]]]
 
         '''
@@ -1891,14 +1887,12 @@ class Data(Container,
     <{{repr}}Data(1, 2, 1, 4): [[[[12, ..., 57]]]]>
     >>> print(e.array)
     [[[[12 15 18 21]]
-
       [[48 51 54 57]]]]
     >>> e = d.sum([-2, -1])
     >>> e
     <{{repr}}Data(1, 2, 1, 1): [[[[66, 210]]]]>
     >>> print(e.array)
     [[[[ 66]]
-
       [[210]]]]
 
         '''
