### 2020-10-02 #2531 @usethedata
Sorry to have been absent here.  I was slammed with end of fiscal year stuff and a reorganization at work.  I could have sworn I opened the review issue for this, but didn't.  This is the pre-review issue.  Let me actually create the review issue and then I'll copy/post the work @bradyrx has already done over into that issue and go from there.
