### 2020-10-02 #2717 @whedon
[REVIEW]: cfdm: A Python reference implementation of the CF data model
**Submitting author:** @davidhassell (<a href="http://orcid.org/0000-0001-5106-7502">David Hassell</a>)
**Repository:** <a href="https://github.com/NCAS-CMS/cfdm" target ="_blank">https://github.com/NCAS-CMS/cfdm</a>
**Version:** 1.8.7.0

@bradyrx @clynnes, please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
