### 2020-10-02 #2717 @usethedata
@bradyrx  comments from the pre-review issue (https://github.com/openjournals/joss-reviews/issues/2531) because I didn't do my job correctly and didn't get the review issue started:

> Hi all, thanks for the patience here. It looks like this is stuck in pre-review? I have time now to do this so I'm just going to post my review here using the checklist I used last time I reviewed for JOSS.
>
>Overall, cfdm looks great and I'm happy to suggest its acceptance to JOSS following some minor corrections I've listed below. Some are nitpicky grammar on the paper, others are some thoughts to make contributions more accessible to early open-source contributors. I was able to use cfdm very quickly with some forecasting data I had saved out to disk. The docs are thorough and give good tutorials. I didn't go too deep into the code base, but after skimming through, it hits the points I care about -- robust testing, leverages inheritance, and has consistently good docstrings.
>
> Conflict of interest
> X I confirm that I have read the JOSS conflict of interest (COI) policy and that: I have no COIs with reviewing this work or that any perceived COIs have been waived by JOSS for the purpose of this review.
>
> Code of Conduct
> X I confirm that I read and will adhere to the JOSS code of conduct.
>
> General checks
> X Repository: Is the source code for this software available at the repository url?
> > Yes. The source code is available and the repo is very well structured for open-source science.
> X License: Does the repository contain a plain-text LICENSE file with the contents of an OSI approved software license?
> X Contribution and authorship: Has the submitting author (@davidhassell) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
> > Yes, it looks like the two main contributors to the code base are the two authors.
>
> Functionality
> X Installation: Does installation proceed as outlined in the documentation?
> X Functionality: Have the functional claims of the software been confirmed?
> > Yes, I loaded in some SubX subseasonal forecasting output I had and was able to follow along with the tutorial and inspect its properties.
> X Performance: If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)
> > Yes, I tried playing around with inplace operations (in memory) per the docs and things were faster in place.
>
> Documentation
> X A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?
> X  Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.
> X Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
> > Really impressed by the thorough tutorial examples on the docs. Well done!
> X Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
> X Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?
> > There's no codecov or coveralls indicator of how much coverage the testing gives, but I can see that there is a thorough amount of testing done under cfdm/test.
> [ ]  Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
> > 1. I think it would be good to add a code of conduct. See Insights/Community'. That's the main thing missing (outside of PR templates). They have some stock ones that work well. Ours at climpred` is based on https://www.contributor-covenant.org/version/1/4/code-of-conduct/.
> > 2. It looks like CONTRIBUTING.md in the main repo is decoupled from the "support" page on the docs. I.e., the docs are missing some of this information. I would also suggest adding more information for first-time or early contributors, since there's a lot of steps that go into contributing to a python package on Github. I have a lightweight personal package with some contribution instructions that you're welcome to pull from: https://esmtools.readthedocs.io/en/stable/contributing.html (see "Preparing Pull Requests").
> > 3. I might change the docs page name from "Support" to something more obvious like "How to Contribute".
Software paper
> X Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?
> X A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?
> [ ]  State of the field: Do the authors describe how this software compares to other commonly-used packages?
> > It is mentioned that cfdm is the only library that is guaranteed to handle every possible type of CF-compliant dataset. I think this leaves out popular high-level packages like iris (https://github.com/SciTools/iris) that have CF-compliant manipulation of Earth System data, and it doesn't look like cfdm is a dependency there (please correct me if I'm wrong). xarray doesn't natively handle CF well, but an extension to it cf-xarray does: https://github.com/xarray-contrib/cf-xarray. So there might be some other packages that can be nodded to that think of the CF data model.

> [ ]  Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?
> > Generally well written. See my bottom notes on small fixes to the paper's grammar. There are some spurious semicolons and commas in a few places throughout. Please double check uses of , and ;.
> X References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax?

> > Some notes on the paper:
> >
> > • Under the summary, the semicolon should be dropped from "They cater for data from model simulations; and from observations, ...". Maybe change the "and" following it to "as well as" for flow.
> > • The sentence following the previous is structured somewhat awkwardly. I would say something like "CF conventions provide a description of the physical meaning of netCDF data variables, as well as their spatial, temporal, and other dimensional properties."
