### 2020-10-02 #2717 @usethedata
And reproducing @davidhassell reply to the above from the pre-review issue

> @bradyrx, @usethedata
>
> Hi. Thanks again for the detailed review. We have taken all of of your points on board, and hope that we can address them with the changes described here.
>
> > There's no codecov or coveralls indicator of how much coverage the testing gives, but I can see that there is a thorough amount of testing done under cfdm/test.
>
> Code coverage is calculated as part of the test suite, and run with the GitHub actions, but is not easy to find. It is currently 87% (e.g.  The "Run test suite and generate a coverage report" section of https://github.com/NCAS-CMS/cfdm/runs/1103183661). We are currently working to make this accessible by adding a new workflow to our GitHub Actions setup which will integrate the Codecov tool to the project and trigger coverage reports against every pull request..
>
> > I think it would be good to add a code of conduct. See Insights/Community'. That's the main thing missing (outside of PR templates). They have some stock ones that work well. Ours at climpred` is based on https://www.contributor-covenant.org/version/1/4/code-of-conduct/.
>
> There is now a code of conduct at https://github.com/NCAS-CMS/cfdm/blob/master/CODE_OF_CONDUCT.md
>
> > It looks like CONTRIBUTING.md in the main repo is decoupled from the "support" page on the docs. I.e., the docs are missing some of this information. I would also suggest adding more information for first-time or early contributors, since there's a lot of steps that go into contributing to a python package on Github. I have a lightweight personal package with some contribution instructions that you're welcome to pull from: https://esmtools.readthedocs.io/en/stable/contributing.html (see "Preparing Pull Requests").
> >
> > I might change the docs page name from "Support" to something more obvious like "How to Contribute".
>
> We have replaced the "Support" section of the docs with a new "Contributing" section: https://github.com/NCAS-CMS/cfdm/blob/master/docs/source/contributing.rst, which draws on the esmtools approach - thanks. This file will get rendered as html at the time of the next release.
>
> The CONTRIBUTING.md file (https://github.com/NCAS-CMS/cfdm/blob/master/CONTRIBUTING.md) has been modified to be consistent with the new docs. Instead of just duplicating the content from the Contributing section of the documentation, we preferred to provide a summary there and link to the docs page for full detail.
>
> > Generally well written. See my bottom notes on small fixes to the paper's grammar. There are some spurious semicolons and commas in a few places throughout. Please double check uses of , and ;.
>
> We have revised the grammar in this light. The new text is in https://github.com/NCAS-CMS/cfdm/blob/master/joss/paper.md, and in the new whedon-generated pdf following this comment. The changes made in this respect, should you wish to isolate them, are contained in the pull request https://github.com/NCAS-CMS/cfdm/pull/76.
>
> > It is mentioned that cfdm is the only library that is guaranteed to handle every possible type of CF-compliant dataset. I think this leaves out popular high-level packages like iris (https://github.com/SciTools/iris) that have CF-compliant manipulation of Earth System data, and it doesn't look like cfdm is a dependency there (please correct me if I'm wrong). xarray doesn't natively handle CF well, but an extension to it cf-xarray does: https://github.com/xarray-contrib/cf-xarray. So there might be some other packages that can be nodded to that think of the CF data model.
>
> I think that our  original text was indeed less than clear on the separation between being able to manipulate all possible CF-netCDF datasets and implementing the CF data model. The former is possible without the latter, and the latter does not guarantee the former. Many libraries (iris, xarray, cf-xarray, ferret, grads, netCDF Java, ...) understand many features of CF, but none of them implement the CF data model, and also none are complete - in the sense that there are some elements of CF that they do not understand. cfdm both implements the CF data model, and can process every all possible CF-netCDF datasets.
>
> Iris is a interesting example, as they state that they implement "a" data model of the CF conventions (my emphasis), but do not state what it is. It is not, however, the accepted CF data model that cfdm implements (for example, coordinate systems in iris are contained within coordinates, rather than the parent iris cube, which the CF data model would require). This is of course not a problem per se, as the CF data model makes no claims on uniqueness, but in any case iris does not yet provide functionality for everything in CF. From a look around the code base and docs, it appears to not be able to deal with simple geometries, netCDF hierarchical groups, discrete sampling geometries, and compression by gathering (for example).
>
> I hope that a change to the text in the "Statement of need" section will be enough to clear up this subtle area:
>
> **(Old text)** "To the authors' knowledge, cfdm and software built on it are currently the only libraries that are guaranteed to be able to handle every possible type of CF-compliant dataset. All others omit facets that are not currently of interest to their particular user communities."
>
> **(New text)** "To the authors' knowledge, cfdm and software built on it are currently the only libraries that fully implement the CF data model and its application to every feature of the CF-netCDF encoding. All others omit facets that are not currently of interest to their particular user communities."
>
> (the new text is also in the latest pdf version).
