### 2020-10-02 #2717 @usethedata
@bradyrx  -- Based on the your comments over in the pre-review, I believe you're satisfied.  If you'd be so kind as to check off the items in your checklist, that would be much appreciated.

@davidhassell -- My apologies again for the delay, and I do need to get a second reviewer, which I'm now working on.
