### 2020-10-04 #2717 @davidhassell
@bradyrx - just to let you know that, further to your comments on code coverage, @sadielbartholomew has now implemented a codecov integration: https://codecov.io/gh/NCAS-CMS/cfdm, also available as a link in the cfdm GitHUb page https://github.com/NCAS-CMS/cfdm.
