### 2020-10-05 a01e924
Fixed merge conflicts
