+++ docs/source/class.rst
@@ -18,7 +18,7 @@ Field construct class
    :nosignatures:
    :toctree: class/
 		 
-   cfdm.Field	              
+   cfdm.Field
 
 Metadata construct classes
 --------------------------
