### 2020-10-05 0850f06
fixed some missing api docs
