### 2020-10-05 c7fe807
fixed some missing api docs
