### 2020-10-05 522f771
fixed some missing api docs
