+++ docs/source/class/cfdm.core.abstract.Container.rst
@@ -22,18 +22,6 @@ Miscellaneous
 
    ~cfdm.core.abstract.Container.copy
 
-Special
--------
-
-.. rubric:: Methods
-
-.. autosummary::
-   :nosignatures:
-   :toctree: ../method/
-   :template: method.rst
-
-   ~cfdm.core.abstract.Container.__deepcopy__
-
 Private
 -------
 
@@ -49,3 +37,29 @@ Private
    ~cfdm.core.abstract.Container._has_component
    ~cfdm.core.abstract.Container._set_component
 
+Special
+-------
+
+.. rubric:: Methods
+
+.. autosummary::
+   :nosignatures:
+   :toctree: ../method/
+   :template: method.rst
+
+   ~cfdm.core.abstract.Container.__deepcopy__
+
+Docstring substitutions
+-----------------------                                 
+                                                        
+.. rubric:: Methods                                     
+                                                        
+.. autosummary::                                        
+   :nosignatures:                                       
+   :toctree: ../method/                                 
+   :template: method.rst                                
+                                                        
+   ~cfdm.core.abstract.Container._docstring_special_substitutions
+   ~cfdm.core.abstract.Container._docstring_substitutions        
+   ~cfdm.core.abstract.Container._docstring_package_depth        
+   ~cfdm.core.abstract.Container._docstring_method_exclusions    
