+++ docs/source/class/cfdm.core.NumpyArray.rst
@@ -36,7 +36,7 @@ Miscellaneous
    
    ~cfdm.core.NumpyArray.copy
 
-Docstring substitutions                           
+Docstring substitutions
 -----------------------                        
                                                
 .. rubric:: Methods                            
