### 2020-10-05 627ca12
fixed some missing api docs
