+++ docs/source/check_docs_api_coverage.py
@@ -95,3 +95,4 @@ if n_undocumented_methods or n_missing_files:
         )
     )
 
+print('All methods are documented')
