### 2020-10-05 9706e7d
fixed some missing api docs
