+++ cfdm/core/__init__.py
@@ -86,5 +86,4 @@ from .abstract import (
     ParametersDomainAncillaries,
 )
 
-
 from .meta import DocstringRewriteMeta
