+++ docs/source/class/cfdm.core.abstract.Coordinate.rst
@@ -1,131 +1,0 @@
-.. currentmodule:: cfdm
-.. default-role:: obj
-
-cfdm.core.abstract.Coordinate
-=============================
-
-----
-
-.. autoclass:: cfdm.core.abstract.Coordinate
-   :no-members:
-   :no-inherited-members:
-   
-Properties
-----------
-
-.. rubric:: Methods
-
-.. autosummary::
-   :nosignatures:
-   :toctree: ../method/
-   :template: method.rst
-      
-   ~cfdm.core.abstract.Coordinate.del_property
-   ~cfdm.core.abstract.Coordinate.get_property
-   ~cfdm.core.abstract.Coordinate.has_property
-   ~cfdm.core.abstract.Coordinate.set_property
-   ~cfdm.core.abstract.Coordinate.properties
-   ~cfdm.core.abstract.Coordinate.clear_properties
-   ~cfdm.core.abstract.Coordinate.set_properties
-
-Data
-----
-
-.. rubric:: Methods
-
-.. autosummary::
-   :nosignatures:
-   :toctree: ../method/
-   :template: method.rst
-
-   ~cfdm.core.abstract.Coordinate.del_data
-   ~cfdm.core.abstract.Coordinate.get_data
-   ~cfdm.core.abstract.Coordinate.has_data
-   ~cfdm.core.abstract.Coordinate.set_data
-   
-.. rubric:: Attributes
-   
-.. autosummary::
-   :nosignatures:
-   :toctree: ../attribute/
-   :template: attribute.rst
-
-   ~cfdm.core.abstract.Coordinate.data
-
-Bounds
-------
-
-.. rubric:: Methods
-
-.. autosummary::
-   :nosignatures:
-   :toctree: ../method/
-   :template: method.rst
-
-   ~cfdm.core.abstract.Coordinate.del_bounds
-   ~cfdm.core.abstract.Coordinate.get_bounds
-   ~cfdm.core.abstract.Coordinate.has_bounds
-   ~cfdm.core.abstract.Coordinate.set_bounds
-   
-.. rubric:: Attributes
-   
-.. autosummary::
-   :nosignatures:
-   :toctree: ../attribute/
-   :template: attribute.rst
-
-   ~cfdm.core.abstract.Coordinate.bounds
-   
-Modification
-------------
-
-.. rubric:: Methods
-	    
-.. autosummary::
-   :nosignatures:
-   :toctree: ../method/
-   :template: method.rst
-
-   ~cfdm.core.abstract.Coordinate.del_data
-   ~cfdm.core.abstract.Coordinate.set_data
-   ~cfdm.core.abstract.Coordinate.del_bounds
-   ~cfdm.core.abstract.Coordinate.set_bounds
-      
-Miscellaneous
--------------
-
-.. rubric:: Methods
-	    
-.. autosummary::
-   :nosignatures:
-   :toctree: ../method/
-   :template: method.rst
-
-   ~cfdm.core.abstract.Coordinate.copy
-
-Special
--------
-
-.. rubric:: Methods
-
-.. autosummary::
-   :nosignatures:
-   :toctree: ../method/
-   :template: method.rst
-
-   ~cfdm.core.abstract.Coordinate.__deepcopy__
-
-Docstring substitutions
------------------------                                 
-                                                        
-.. rubric:: Methods                                     
-                                                        
-.. autosummary::                                        
-   :nosignatures:                                       
-   :toctree: ../method/                                 
-   :template: method.rst                                
-                                                        
-   ~cfdm.core.abstract.Coordinate._docstring_special_substitutions
-   ~cfdm.core.abstract.Coordinate._docstring_substitutions        
[…]
