### 2020-10-05 ab1f99f
Actions: update now-misleading comment in workflow
