+++ cfdm/functions.py
@@ -564,64 +564,3 @@ def abspath(filename):
         return filename
 
     return os.path.abspath(filename)
-
-
-def unique_domains(fields, copy=True):
-    '''Return the unique domains used by the field constructs.
-
-    .. versionadded:: (cfdm) 1.8.7.0
-
-    :Parameters:
-
-        fields: seqeunce of `Field`
-            The fields to compared. May be an empty sequence.
-
-        copy: `bool`, optional
-            If False then do not copy returned domains. By default the
-            returned domains are deep copies from their parent field
-            constructs.
-
-    :Returns:
-
-        `list` of `Domain`
-            The unique domains. May be an empty list.
-
-    **Examples:**
-
-    >>> f = cfdm.example_fields(0)
-    >>> g = cfdm.example_fields(1)
-    >>> f
-    <Field: specific_humidity(latitude(5), longitude(8)) 1>
-    >>> g
-    <Field: air_temperature(atmosphere_hybrid_height_coordinate(1), grid_latitude(10), grid_longitude(9)) K>
-    >>> cfdm.unique_domains([f, f, g])
-    [<Domain: {1, 5, 8}>, <Domain: {1, 1, 9, 10}>]
-
-    '''
-    if not fields:
-        return []
-
-    domain = fields[0].domain
-    if copy:
-        domain = domain.copy()
-
-    out = [domain]
-
-    for f in fields[1:]:
-        domain = f.domain
-
-        found_new_domain = False
-        for d in out:
-            if not d.equals(domain):
-                found_new_domain = True
-                break
-        # --- End: for
-
-        if found_new_domain:
-            if copy:
-                domain = domain.copy()
-
-            out.append(domain)
-    # --- End: for
-
-    return out
