+++ cfdm/test/test_dsg.py
@@ -33,31 +33,28 @@ atexit.register(_remove_tmpfiles)
 
 
 class DSGTest(unittest.TestCase):
-    def setUp(self):
-        # Disable log messages to silence expected warnings
-        cfdm.log_level('DISABLE')
-        # Note: to enable all messages for given methods, lines or
-        # calls (those without a 'verbose' option to do the same)
-        # e.g. to debug them, wrap them (for methods, start-to-end
-        # internally) as follows:
-        #
-        # cfdm.log_level('DEBUG')
-        # < ... test code ... >
-        # cfdm.log_level('DISABLE')
+    contiguous = os.path.join(
+        os.path.dirname(os.path.abspath(__file__)),
+        'DSG_timeSeries_contiguous.nc')
+    indexed = os.path.join(
+        os.path.dirname(os.path.abspath(__file__)),
+        'DSG_timeSeries_indexed.nc')
+    indexed_contiguous = os.path.join(
+        os.path.dirname(os.path.abspath(__file__)),
+        'DSG_timeSeriesProfile_indexed_contiguous.nc')
 
-        self.contiguous = 'DSG_timeSeries_contiguous.nc'
-        self.indexed = 'DSG_timeSeries_indexed.nc'
-        self.indexed_contiguous = 'DSG_timeSeriesProfile_indexed_contiguous.nc'
+    c = cfdm.read(contiguous)
+    i = cfdm.read(indexed)
+    ic = cfdm.read(indexed_contiguous)
 
-        a = numpy.ma.masked_all((4, 9), dtype=float)
-        a[0, 0:3] = [0.0, 1.0, 2.0]
-        a[1, 0:7] = [1.0, 11.0, 21.0, 31.0, 41.0, 51.0, 61.0]
-        a[2, 0:5] = [2.0, 102.0, 202.0, 302.0, 402.0]
-        a[3, 0:9] = [3.0, 1003.0, 2003.0, 3003.0, 4003.0,
-                     5003.0, 6003.0, 7003.0, 8003.0]
-        self.a = a
+    a = numpy.ma.masked_all((4, 9), dtype=float)
+    a[0, 0:3] = [0.0, 1.0, 2.0]
+    a[1, 0:7] = [1.0, 11.0, 21.0, 31.0, 41.0, 51.0, 61.0]
+    a[2, 0:5] = [2.0, 102.0, 202.0, 302.0, 402.0]
+    a[3, 0:9] = [3.0, 1003.0, 2003.0, 3003.0, 4003.0,
+                 5003.0, 6003.0, 7003.0, 8003.0]
 
-        b = numpy.array(
+    b = numpy.array(
             [[[20.7, -99, -99, -99],
               [10.1, 11.8, 18.2, -99],
               [11.0, 11.8, -99, -99],
@@ -140,16 +137,127 @@ class DSGTest(unittest.TestCase):
               [-99, -99, -99, -99]]]
         )
 
-        b = numpy.ma.where(b == -99, numpy.ma.masked, b)
-        self.b = b
+    b = numpy.ma.where(b == -99, numpy.ma.masked, b)
 
-        self.test_only = []
+    test_only = []
+
+    def setUp(self):
+        # Disable log messages to silence expected warnings
+        cfdm.log_level('DISABLE')
+        # Note: to enable all messages for given methods, lines or
+        # calls (those without a 'verbose' option to do the same)
+        # e.g. to debug them, wrap them (for methods, start-to-end
+        # internally) as follows:
+        #
+        # cfdm.log_level('DEBUG')
+        # < ... test code ... >
+        # cfdm.log_level('DISABLE')
+
+#        self.contiguous = 'DSG_timeSeries_contiguous.nc'
[…]
