+++ cfdm/test/test_functions.py
@@ -216,7 +216,7 @@ class FunctionsTest(unittest.TestCase):
             cfdm.write(f, temp_file)
             g = cfdm.read(temp_file, verbose=1)
 
-            self.assertEqual(len(g), 1, g)
+            self.assertEqual(len(g), 1)
             self.assertTrue(f.equals(g[0], verbose=3), 'n={}'.format(n))
 
         with self.assertRaises(Exception):
