### 2020-10-07 e32b1f1
Remove support for Python 3.5 (end-of-life as of Sept '20)
