+++ setup.py
@@ -161,6 +161,6 @@ setup(
         'cfdm.test',
     ],
     scripts=['scripts/cfdump'],
-    python_requires='>=3.5',
+    python_requires='>=3.6',
     install_requires=install_requires,
 )
