### 2020-10-08 83d4725
trap for incompatible bounds units from cfdm.read
