### 2020-10-09 #2717 @usethedata
@davidhassell I show 1.8.6 as the current version here and matching what's current in your repository.  If that's correct, then please go ahead and make a Zenodo archive of this current version and post the DOI here.  If you're doing an update of the version to reflect changes made in this review, then update your repository to show the current version, make a Zenodo archive of that.  Let me know the current version number and the Zenodo DOI.  Thanks.

I have no further changes to make in the paper, and I've validated the references.
