+++ cfdm/core/__init__.py
@@ -3,7 +3,7 @@
 '''
 
 __author__ = 'David Hassell'
-__date__ = '2020-10-??'
+__date__ = '2020-10-09'
 __cf_version__ = '1.8'
 __version__ = '1.8.7.0'
 
