+++ MANIFEST.in
@@ -2,4 +2,5 @@ include Changelog.rst
 include README.md
 include requirements.txt
 include LICENSE
-include CONTRIBUTING.md 
+include CONTRIBUTING.md
+include CODE_OF_CONDUCT.md
