### 2020-10-09 #2717 @davidhassell
@usethedata Thanks, that's great. We have just released version 1.8.7.0 with the changes discussed here:
```
10.5281/zenodo.4075077
```
