### 2020-10-09 #2717 @davidhassell
Hi @usethedata, I've just spotted that I have missed out an author in one of the references - is too late to add it in?

Many thanks,
David
