### 2020-10-09 #2717 @sadielbartholomew
> I've just spotted that I have missed out an author in one of the references - is too late to add it in?

> > Go ahead and add.

Hi @usethedata, just for your information, David asked me offline to keep an eye on comments here and to if permitted to do this change since I work a late schedule so I am still working at this hour and he has probably started his weekend.

Let me just add a one-line commit to add in the extra author, as mentioned above, then I will comment here as after that to confirm it is done. Then everything is ready-to-go from our perspective. Thanks.
