+++ joss/paper.bib
@@ -68,7 +68,7 @@
 }
 
 @software{Hassell:2020,
-  author       = {{Hassell}, D.},
+  author       = {{Hassell}, D. and {Bartholomew}, S. L.},
   title        = {{{cfdm}}: A {{Python}} reference implementation of the {{CF}} data model},
   month        = jul,
   year         = 2020,
