### 2020-10-09 #2717 @sadielbartholomew
Hi @usethedata, thanks David actually was still online and beat me to it so the change has now been made in this one new commit: NCAS-CMS/cfdm@0f8cbeafa32c77c5fc6e2267b97d22d3d2b27b1f.

I'm just wondering if it would be best we update the Zenodo version to the one just released, too, i.e. make this precise change:

```diff
diff --git a/joss/paper.bib b/joss/paper.bib
index 86b0437cc..02241916b 100644
--- a/joss/paper.bib
+++ b/joss/paper.bib
@@ -68,15 +68,15 @@
 }

 @software{Hassell:2020,
   title        = {{{cfdm}}: A {{Python}} reference implementation of the {{CF}} data model},
-  month        = jul,
+  month        = oct,
   year         = 2020,
   publisher    = {Zenodo},
-  version      = {1.8.6.0},
-  doi          = {10.5281/zenodo.3894524},
+  version      = {1.8.7.0},
+  doi          = {10.5281/zenodo.4075077},
   urldate      = {2020-07-27},
-  url          = {https://doi.org/10.5281/zenodo.3894524}
+  url          = {https://doi.org/10.5281/zenodo.4075077}
 }
```

Or shall we just leave it?
