### 2020-10-09 f720202
JOSS paper: update Zenodo software reference
