+++ joss/paper.bib
@@ -70,13 +70,13 @@
 @software{Hassell:2020,
   author       = {{Hassell}, D. and {Bartholomew}, S. L.},
   title        = {{{cfdm}}: A {{Python}} reference implementation of the {{CF}} data model},
-  month        = jul,
+  month        = oct,
   year         = 2020,
   publisher    = {Zenodo},
-  version      = {1.8.6.0},
-  doi          = {10.5281/zenodo.3894524},
+  version      = {1.8.7.0},
+  doi          = {10.5281/zenodo.4075077},
   urldate      = {2020-07-27},
-  url          = {https://doi.org/10.5281/zenodo.3894524}
+  url          = {https://doi.org/10.5281/zenodo.4075077}
 }
 
 @article{Juckes:2020,
