### 2020-10-09 #2717 @sadielbartholomew
Please ignore the above, we decided that change should be made so I have pushed up a final commit to update the Zenodo version and corresponding metadata from 1.8.6.0 to 1.8.7.0, as per the tagged version on the paper: NCAS-CMS/cfdm@f720202ce453b9e67643a114f1cddfa8df887ab9.

Thanks for your patience. This is now ready to go as far as we are concerned.
