### 2020-10-10 #2717 @kthyng
Hi! I can take over from here. I do have a question though. It's unusual for a reviewer to not say anything in a review issue — did some part of the review from @clynnes occur offline or just not linked in here so I don't see it?  Was there confirmation that the reviewer is satisfied with proceeding?
