### 2020-10-10 #2717 @kthyng
@davidhassell Please edit the metadata for title and authors for your zenodo archive so the title and author list exactly match your JOSS paper, unless you prefer not to have it match.
