### 2020-10-10 #2717 @kthyng
The version has been updated, and the paper looks good to me.
