### 2020-10-10 #2717 @usethedata
@kthyng Apologies for not documenting the discussion @clynnes and I had on the subject.  He didn't see anything that needed to be changed beyond what was already addressed.  That's a new reviewer item that I should have handled better.
