### 2020-10-10 #2717 @davidhassell
Thank you @kthyng  and @usethedata. I have updated the zenodo archive to match the title and authors in the JOSS paper (https://doi.org/10.5281/zenodo.4075077).
